refactor(app): split chat runtime and transport
