+++ apps/screenpipe-app-tauri/components/standalone-chat.tsx
@@ -5,15 +5,7 @@
 
 import * as React from "react";
 import { useState, useRef, useEffect, useCallback, useMemo } from "react";
-import { listen, type UnlistenFn } from "@tauri-apps/api/event";
-import {
-  mountAgentEventBus,
-  registerForeground,
-  onTerminated as onAgentTerminated,
-  onEvicted as onAgentEvicted,
-} from "@/lib/events/bus";
-import { pipeSessionId } from "@/lib/events/types";
-import { useSettings, ChatConversation } from "@/lib/hooks/use-settings";
+import { useSettings } from "@/lib/hooks/use-settings";
 import { cn } from "@/lib/utils";
 import { SchedulePromptDialog } from "@/components/chat/schedule-prompt-dialog";
 import { BrowserSidebar } from "@/components/browser-sidebar";
@@ -21,17 +13,13 @@ import { toast } from "@/components/ui/use-toast";
 import { AIPreset, PiQueuedPrompt } from "@/lib/utils/tauri";
 // OpenAI SDK no longer used directly — all providers route through Pi agent
 import posthog from "posthog-js";
-import { open as openUrl } from "@tauri-apps/plugin-shell";
 import {
   docsToPromptText,
 } from "@/lib/pi/extract-document";
 import { commands } from "@/lib/utils/tauri";
-import { emit } from "@tauri-apps/api/event";
 import { useChatConversations } from "@/components/hooks/use-chat-conversations";
 import { useChatStore } from "@/lib/stores/chat-store";
 import { useFeedbackStore } from "@/lib/stores/feedback-store";
-import { handlePiEvent, statusForEvent } from "@/lib/stores/pi-event-router";
-import { deriveFallbackConversationTitle } from "@/lib/utils/chat-title";
 import { getCurrentWindow } from "@tauri-apps/api/window";
 import { usePlatform } from "@/lib/hooks/use-platform";
 import { useHardcodedTiles } from "@/lib/hooks/use-hardcoded-tiles";
@@ -39,28 +27,16 @@ import { useIsFullscreen } from "@/lib/hooks/use-is-fullscreen";
 import { useChatFilePreview } from "@/lib/hooks/use-chat-file-preview";
 import { useSqlAutocomplete, useTagAutocomplete } from "@/lib/hooks/use-sql-autocomplete";
 import { homeDir, join } from "@tauri-apps/api/path";
-import { useTimelineStore } from "@/lib/hooks/use-timeline-store";
 import {
   buildAppMentionSuggestions,
   buildTagMentionSuggestions,
   normalizeAppTag,
-  extractConversationHistorySyncUserText,
-  isInjectedTitleSourcePrompt,
 } from "@/lib/chat-utils";
 import { useAutoSuggestions } from "@/lib/hooks/use-auto-suggestions";
 import {
-  buildDailyLimitMessage,
-  classifyQuotaError,
-  buildRateLimitMessage,
-  parseRateLimitWaitSeconds,
-  PI_MAX_RATE_LIMIT_RETRIES,
-} from "@/lib/chat/quota-errors";
-import {
   buildInvalidatedAuthTokenMessage,
-  isInvalidatedAuthTokenError,
 } from "@/lib/chat/auth-errors";
 import {
-  buildNoResponseMessage,
   buildProviderErrorMessage,
   preflightChatProvider,
 } from "@/lib/chat/provider-errors";
@@ -77,18 +53,12 @@ import {
   computeChatCitationPlan,
 } from "@/lib/source-citations";
 import {
-  imageDataUrlsFromPiContent,
   imageDataUrlsToPiImages,
 } from "@/lib/chat/image-content";
 import {
   queuedPreviewForText,
 } from "@/lib/chat/queued-display";
 import { withConversationHistory } from "@/lib/chat/conversation-history";
-import {
-  formatSearchToolError,
-  formatSearchToolResults,
-  type SearchToolResult,
-} from "@/lib/chat/search-tool-results";
 import { INTEGRATION_ICON_KEYS } from "@/components/settings/connections-section";
 import { ImageViewerDialog } from "@/components/chat/standalone/image-viewer-dialog";
 import { StandaloneChatHeader } from "@/components/chat/standalone/standalone-chat-header";
@@ -102,15 +72,23 @@ import { usePiChatState } from "@/components/chat/standalone/hooks/use-pi-chat-s
 import { useChatQueue } from "@/components/chat/standalone/hooks/use-chat-queue";
 import { useChatStreamRender } from "@/components/chat/standalone/hooks/use-chat-stream-render";
 import { useChatConversationEvents } from "@/components/chat/standalone/hooks/use-chat-conversation-events";
-import { useChatPipeWatch } from "@/components/chat/standalone/hooks/use-chat-pipe-watch";
 import { useChatMessageActions } from "@/components/chat/standalone/hooks/use-chat-message-actions";
 import { useChatExportMenu } from "@/components/chat/standalone/hooks/use-chat-export-menu";
+import { useChatSessionRuntime } from "@/components/chat/standalone/hooks/use-chat-session-runtime";
+import { usePiForegroundEvents } from "@/components/chat/standalone/hooks/use-pi-foreground-events";
+import { usePiSendTransport } from "@/components/chat/standalone/hooks/use-pi-send-transport";
+import { useChatPanelEffects } from "@/components/chat/standalone/hooks/use-chat-panel-effects";
 import {
   useChatComposerShell,
   useChatComposerShellActions,
 } from "@/components/chat/standalone/hooks/use-chat-composer-shell";
 import { useChatExternalEvents } from "@/components/chat/standalone/hooks/use-chat-external-events";
 import { usePiSessionLifecycle } from "@/components/chat/standalone/hooks/use-pi-session-lifecycle";
+import { useChatTurnIntents } from "@/components/chat/standalone/hooks/use-chat-turn-intents";
+import { useChatSteering } from "@/components/chat/standalone/hooks/use-chat-steering";
+import { useNextTurnAttachments } from "@/components/chat/standalone/hooks/use-next-turn-attachments";
+import { useChatComposerDraftSync } from "@/components/chat/standalone/hooks/use-chat-composer-draft-sync";
+import { usePipeWatchSession } from "@/components/chat/standalone/hooks/use-pipe-watch-session";
 import { useChatTemplateSettings } from "@/components/chat/standalone/hooks/use-chat-template-settings";
 import { useTryInChatEvent } from "@/components/chat/standalone/hooks/use-try-in-chat-event";
 import {
@@ -121,14 +99,8 @@ import {
   usePipeGenerationCompletion,
 } from "@/components/chat/standalone/hooks/use-chat-window-events";
 import type {
-  ChatAttachment,
-  ContentBlock,
   Message,
-  OptimisticSteerPayload,
   PendingSteerBatchItem,
-  QueuedDisplayPayload,
-  ToolCall,
-  TurnIntentRecord,
 } from "@/lib/chat/types";
 
 // Session ID is per-conversation — set on mount (new conv) and updated on load/new.
@@ -137,7 +109,6 @@ import type {
 const APP_SUGGESTION_LIMIT = 10;
 const TAG_SUGGESTION_LIMIT = 10;
 const STREAM_RENDER_THROTTLE_MS = 80;
-const POST_STREAM_SIDE_EFFECT_DELAY_MS = 1_500;
 const CHAT_RAIL_CLASS = "max-w-4xl mx-auto w-full";
 
 
@@ -158,8 +129,6 @@ const STATIC_MENTION_SUGGESTIONS: MentionSuggestion[] = [
  * return a user-facing message appropriate to their actual subscription tier.
  */
 
-
-const TURN_INTENT_LEDGER_TTL_MS = 10 * 60 * 1000;
 
 /**
  * Title + actions for the current chat. Click → menu with Rename
@@ -414,21 +383,6 @@ export function StandaloneChat({
     setShowMentionDropdown,
     setMentionFilter,
   });
-  // Single-shot stash of the attachment metadata for the NEXT user message
-  // about to be created by sendPiMessage / enqueuePiMessage. sendMessage
-  // populates this just before dispatching; the message-creation sites
-  // read-and-clear it. We don't carry attachments via the existing
-  // displayLabel string because that's already overloaded (it's the
-  // collapsed bubble label) and because each send path creates its own
-  // Message object — a ref keeps the surface area tiny vs. plumbing a
-  // new param through five call sites.
-  const pendingAttachmentsRef = useRef<ChatAttachment[]>([]);
-  function consumePendingAttachments(): ChatAttachment[] | undefined {
-    const list = pendingAttachmentsRef.current;
-    if (!list.length) return undefined;
-    pendingAttachmentsRef.current = [];
-    return list;
-  }
   const steerShortcutInFlightRef = useRef(false);
 
   const {
@@ -441,12 +395,6 @@ export function StandaloneChat({
     piStreamingTextRef,
     piMessageIdRef,
     piContentBlocksRef,
-    pendingNextPiUserIntentRef,
-    pendingNextPiUserDisplayRef,
-    optimisticSteerRef,
-    turnIntentLedgerRef,
-    pendingSteerBatchRef,
-    pendingSteerFlushInFlightRef,
     piLastErrorRef,
     invalidatedAuthHandledRef,
     piStartInFlightRef,
@@ -469,6 +417,13 @@ export function StandaloneChat({
     piSessionIdRef,
     piRunningConfigRef,
   } = usePiChatState();
+  const {
+    optimisticSteerRef,
+    pendingNextPiUserDisplayRef,
+    pendingNextPiUserIntentRef,
+    pendingSteerBatchRef,
+    pendingSteerFlushInFlightRef,
+  } = useChatSteering();
 
   const {
     cancelStreamingMessageRender,
@@ -482,7 +437,14 @@ export function StandaloneChat({
     streamRenderThrottleMs: STREAM_RENDER_THROTTLE_MS,
   });
 
-  const normalizeTurnIntentText = (value: string) => value.replace(/\s+/g, " ").trim();
+  const {
+    findTurnIntentForUserStart,
+    markTurnIntentConsumed,
+    registerTurnIntent,
+    removeTurnIntent,
+    turnIntentLedgerRef,
+    turnIntentTextValuesMatch,
+  } = useChatTurnIntents();
 
   useEffect(() => {
     if (settings.user?.token) {
@@ -523,70 +485,6 @@ export function StandaloneChat({
     }
   }, [updateSettings]);
 
-  const turnIntentTextValuesMatch = (leftValue: string, rightValue: string) => {
-    const left = normalizeTurnIntentText(leftValue);
-    const right = normalizeTurnIntentText(rightValue);
-    if (!left || !right) return false;
-    return left === right;
-  };
-
-  const turnIntentMatchesText = (record: TurnIntentRecord, text: string) => {
-    return turnIntentTextValuesMatch(record.content, text) || turnIntentTextValuesMatch(record.preview, text);
-  };
-
-  const pruneTurnIntentLedger = () => {
-    const cutoff = Date.now() - TURN_INTENT_LEDGER_TTL_MS;
-    turnIntentLedgerRef.current = turnIntentLedgerRef.current.filter((record) => record.createdAt >= cutoff);
-  };
-
-  const registerTurnIntent = (record: TurnIntentRecord) => {
-    pruneTurnIntentLedger();
-    turnIntentLedgerRef.current = [
-      ...turnIntentLedgerRef.current.filter((item) => item.id !== record.id),
-      record,
-    ];
-  };
-
-  const removeTurnIntent = (id: string) => {
-    turnIntentLedgerRef.current = turnIntentLedgerRef.current.filter((record) => record.id !== id);
-  };
-
-  const findTurnIntentForUserStart = (
-    sessionId: string | null | undefined,
-    text: string,
-    display?: QueuedDisplayPayload | null,
-  ): TurnIntentRecord | null => {
-    if (!sessionId) return null;
-    pruneTurnIntentLedger();
-    const sessionTurnIntents = turnIntentLedgerRef.current.filter((record) => record.sessionId === sessionId);
-    const hasIncomingText = Boolean(normalizeTurnIntentText(text));
-    const displayPreviewMatchesIncoming = display?.preview
-      ? turnIntentTextValuesMatch(display.preview, text)
-      : false;
-    const canUseDisplayIdentity = Boolean(display && (!hasIncomingText || displayPreviewMatchesIncoming));
-    const recordMatchesIncoming = (record: TurnIntentRecord) =>
-      turnIntentMatchesText(record, text) ||
-      (displayPreviewMatchesIncoming && turnIntentMatchesText(record, display?.preview ?? ""));
-
-    const byDisplayId = canUseDisplayIdentity && display?.turnIntentId
-      ? sessionTurnIntents.find((record) => record.id === display.turnIntentId)
-      : null;
-    if (byDisplayId && recordMatchesIncoming(byDisplayId)) return byDisplayId;
-
-    const byOptimisticUser = canUseDisplayIdentity && display?.optimisticUserId
-      ? sessionTurnIntents.find((record) => record.displayedUserId === display.optimisticUserId)
-      : null;
-    if (byOptimisticUser && recordMatchesIncoming(byOptimisticUser)) return byOptimisticUser;
-
-    return sessionTurnIntents.find((record) => turnIntentMatchesText(record, text)) ?? null;
-  };
-
-  const markTurnIntentConsumed = (id: string, assistantId: string) => {
-    turnIntentLedgerRef.current = turnIntentLedgerRef.current.map((record) =>
-      record.id === id ? { ...record, consumedAssistantId: assistantId } : record
-    );
-  };
-
   const lastUserMessageRef = useRef<string>("");
 
   // Ref to sendMessage so useEffect callbacks can call it without stale closures
@@ -601,17 +499,16 @@ export function StandaloneChat({
   // (it closes over local state setters and refs); routing through a
   // ref avoids an expensive re-extraction.
   const handleAgentEventDataRef = useRef<((data: any) => void) | null>(null);
-  // True until the component unmounts. Used by bus handlers to avoid
-  // touching React state after unmount; equivalent to the per-effect
-  // `mounted` flag but visible across all useEffect boundaries.
-  const mountedRef = useRef(true);
-
   // Chat history state. Initialised to the same uuid as
   // `piSessionIdRef` so the foreground bus key matches Pi's emitted
   // sessionId from message 0 — see comment above piSessionIdRef.
   const [conversationId, setConversationId] = useState<string | null>(
     initialSessionIdRef.current,
   );
+  const {
+    consumePendingAttachments,
+    stagePendingAttachments,
+  } = useNextTurnAttachments(conversationId);
   const { filePreview, openFilePreview, closeFilePreview } =
     useChatFilePreview(conversationId);
   const currentQueueSessionId = conversationId ?? piSessionIdRef.current;
@@ -671,45 +568,15 @@ export function StandaloneChat({
     window.requestAnimationFrame(findAndFocus);
   }, [markUserScrolledUp]);
 
-  // Clear the connection chip whenever the active conversation changes (new chat or history switch).
-  useEffect(() => { clearConnectionChip(); }, [conversationId, clearConnectionChip]);
-
-  useEffect(() => {
-    void refreshConnectionState();
-  }, [conversationId, refreshConnectionState]);
-
-  // Drop any single-shot attachment metadata stashed for the previous
-  // chat's next-send when the user navigates away. Without this, a user
-  // who staged an attachment in chat A and switched to chat B before
-  // sending would have A's attachment metadata silently ride along on
-  // B's next message. Pairs with the composer-state clear inside
-  // loadConversation / startNewConversation.
-  useEffect(() => {
-    pendingAttachmentsRef.current = [];
-  }, [conversationId]);
-
-  // Mirror the live composer into the chat store so the CURRENT chat
-  // always has an up-to-date draft snapshot. This covers the case
-  // where the user closes/hides the panel (or quits Pi without
-  // switching first) — the draft survives because it's in the store,
-  // not just in React state. It also handles the "draft cleared after
-  // send" case for free: when sendMessage calls setInput("") etc.,
-  // the next mirror tick writes an empty draft, which the store action
-  // treats as "drop draft entirely". Debounced so per-keystroke writes
-  // don't churn the store. Skip when conversationId is null (brand-new
-  // chat with no session record yet — setComposerDraft would no-op).
-  useEffect(() => {
-    if (!conversationId) return;
-    const t = setTimeout(() => {
-      useChatStore.getState().actions.setComposerDraft(conversationId, {
-        input,
-        pastedImages,
-        attachedDocs,
-        pendingDocs,
-      });
-    }, 250);
-    return () => clearTimeout(t);
-  }, [conversationId, input, pastedImages, attachedDocs, pendingDocs]);
+  useChatComposerDraftSync({
+    conversationId,
+    input,
+    pastedImages,
+    attachedDocs,
+    pendingDocs,
+    clearConnectionChip,
+    refreshConnectionState,
+  });
 
   // Chat conversations — stored as individual JSON files in ~/.screenpipe/chats/
   const {
@@ -761,42 +628,6 @@ export function StandaloneChat({
   loadConversationRef.current = loadConversation;
   startNewConversationRef.current = startNewConversation;
 
-  const {
-    handleKeyDown,
-    handlePaste,
-    handleSubmit,
-  } = useChatComposerShellActions({
-    input,
-    setInput,
-    inputRef,
-    connectionChip,
-    setConnectionChip,
-    isMac,
-    isComposing,
-    mentions: {
-      isOpen: showMentionDropdown,
-      selectedIndex: selectedMentionIndex,
-      suggestions: filteredMentions,
-    },
-    mentionActions: {
-      close: () => setShowMentionDropdown(false),
-      selectNext: () => setSelectedMentionIndex((index) => Math.min(index + 1, filteredMentions.length - 1)),
-      selectPrevious: () => setSelectedMentionIndex((index) => Math.max(index - 1, 0)),
-      insert: insertMention,
-    },
-    pastedImages,
-    pendingDocsRef,
-    attachedDocsRef,
-    queuedPrompts,
-    steerShortcutInFlightRef,
-    isKnownConnectionId: (id) => INTEGRATION_ICON_KEYS.has(id),
-    handlePastedFiles,
-    attachPastedText,
-    sendMessage,
-    steerMessage,
-    steerQueuedPrompt,
-  });
-
   usePipeGenerationCompletion({ isLoading });
   useChatPrefillListener({
     setIsPreparingPrefill,
@@ -843,239 +674,31 @@ export function StandaloneChat({
     setActivePreset,
   });
 
-  // Component-lifetime guard for bus handlers that fire across the
-  // longer-lived useEffects (terminated, foreground registrations).
-  // Useful because the panel's per-effect `mounted` flags are scoped
-  // to their own effects.
-  useEffect(() => {
-    mountedRef.current = true;
-    return () => {
-      mountedRef.current = false;
-    };
-  }, []);
-
-  // Foreground registration on the agent-event bus. Switches with
-  // `conversationId` so the bus always knows exactly one panel owns
-  // events for the current chat. The router's exclusive routing means
-  // we don't have to filter by sessionId in the handler — the bus
-  // delivers only events whose envelope sessionId matches the
-  // registration key.
-  //
-  // This is also where pipe-watch sessions register: initWatch swaps
-  // conversationId to a `pipe:<name>:<execId>` id, so this effect
-  // re-runs and registers the panel as the foreground owner of pipe
-  // stdout. Switching to a chat unregisters the pipe foreground (via
-  // the cleanup) and registers the chat — pipe events naturally stop
-  // reaching the panel and start hitting the pipe-run-recorder
-  // instead, which is what we want.
-  // Pipe-watch sessions don't register foreground — pipe-watch-writer
-  // is the sole writer for them, panel mirrors store messages below.
-  // We grab `kind` synchronously here (not via the Zustand selector) so
-  // the effect re-runs on conversationId change without an extra render
-  // cycle that could miss the foreground registration window for chats.
-  useEffect(() => {
-    if (!conversationId) return;
-    const kind = useChatStore.getState().sessions[conversationId]?.kind;
-    if (kind === "pipe-watch") return;
-    let cancelled = false;
-    let off: (() => void) | null = null;
-    (async () => {
-      await mountAgentEventBus();
-      if (cancelled) return;
-      off = registerForeground(conversationId, (envelope) => {
-        if (envelope.sessionId !== piSessionIdRef.current) {
-          // The user switched chats, but the old foreground registration
-          // has not cleaned up yet. Hand the event to the background
-          // router so the old session can accrue unread state normally.
-          void handlePiEvent(envelope);
-          return;
-        }
-        if (!mountedRef.current) return;
-        handleAgentEventDataRef.current?.(envelope.event);
-      });
-      // E2E seam: agent_event delivery to the panel is gated on this
-      // foreground registration completing. Tests can wait on this
-      // signal before emitting events; without it, deltas race the
-      // registration window and go to the default router, which
-      // early-returns for `store.currentId === sid`, silently dropping
-      // them. Cleared in the cleanup below so successive switches don't
-      // see a stale id.
-      if (typeof window !== "undefined") {
-        (window as any).__e2eForegroundReady = conversationId;
-      }
-    })();
-    return () => {
-      cancelled = true;
-      try { off?.(); } catch { /* ignore */ }
-      if (typeof window !== "undefined") {
-        if ((window as any).__e2eForegroundReady === conversationId) {
-          (window as any).__e2eForegroundReady = null;
-        }
-      }
-    };
-  }, [conversationId]);
-
-  // Mirror chat-store messages into local React state when the panel is
-  // showing a pipe-watch session. The writer is the source of truth;
-  // this hook makes the existing render path (which reads `messages`)
-  // pick up writer updates without forking the rendering code.
-  const pipeWatchMessages = useChatStore((s) =>
-    conversationId && s.sessions[conversationId]?.kind === "pipe-watch"
-      ? s.sessions[conversationId]?.messages
-      : undefined,
-  );
-  useEffect(() => {
-    if (!pipeWatchMessages) return;
-    setMessages(pipeWatchMessages as any);
-  }, [pipeWatchMessages, setMessages]);
-
-  // Mirror isLoading / isStreaming from the store for pipe-watch
-  // sessions. Without this the panel's "writing…" indicator strands
-  // forever once the pipe finishes — the writer flips the flags in the
-  // store on agent_end, but the panel's local React state was set to
-  // true at initWatch and never gets cleared (no foreground bus
-  // registration → no panel-side terminal handler runs).
-  // Two scalar selectors instead of one returning {isLoading,isStreaming}
-  // — Zustand's shallow-equal would re-render every store mutation if
-  // the selector built a fresh object each call.
-  const pipeWatchIsLoading = useChatStore((s) => {
-    if (!conversationId) return undefined;
-    const sess = s.sessions[conversationId];
-    if (sess?.kind !== "pipe-watch") return undefined;
-    return !!sess.isLoading;
-  });
-  const pipeWatchIsStreaming = useChatStore((s) => {
-    if (!conversationId) return undefined;
-    const sess = s.sessions[conversationId];
-    if (sess?.kind !== "pipe-watch") return undefined;
-    return !!sess.isStreaming;
-  });
-  useEffect(() => {
-    if (pipeWatchIsLoading !== undefined) setIsLoading(pipeWatchIsLoading);
-    if (pipeWatchIsStreaming !== undefined) setIsStreaming(pipeWatchIsStreaming);
-  }, [pipeWatchIsLoading, pipeWatchIsStreaming]);
-
-  // Self-heal a stuck "writing…" indicator on regular chat sessions.
-  // The router (background) and the panel's foreground listener both set
-  // store.isStreaming/isLoading to false on agent_end. Local React
-  // state is also cleared by the foreground listener — but if that
-  // listener missed the event for any reason (race during section
-  // change, mountedRef brief flicker, bus dispatch landing on default
-  // instead of foreground), local can drift `true` while the store
-  // says `false`. Returning to the panel from Settings then shows
-  // "writing…" forever even though the turn is over.
-  // One-way: only sync false→false. The forward direction (true) is
-  // owned by the panel itself when the user sends, so we never want to
-  // reach in and turn the indicator off mid-typing.
-  const storeChatIsStreaming = useChatStore((s) => {
-    if (!conversationId) return undefined;
-    const sess = s.sessions[conversationId];
-    if (!sess || sess.kind === "pipe-watch") return undefined;
-    return !!sess.isStreaming;
-  });
-  const storeChatIsLoading = useChatStore((s) => {
-    if (!conversationId) return undefined;
-    const sess = s.sessions[conversationId];
-    if (!sess || sess.kind === "pipe-watch") return undefined;
-    return !!sess.isLoading;
-  });
-  const currentStreamingMessageId = useChatStore((s) => {
-    if (!conversationId) return null;
-    return s.sessions[conversationId]?.streamingMessageId ?? null;
-  });
-  useEffect(() => {
-    if (storeChatIsStreaming === false) setIsStreaming(false);
-    if (storeChatIsLoading === false) setIsLoading(false);
-  }, [storeChatIsStreaming, storeChatIsLoading]);
-
-  useEffect(() => {
-    isStreamingRef.current = isStreaming;
-  }, [isStreaming]);
-
-  useEffect(() => {
-    isLoadingRef.current = isLoading;
-  }, [isLoading]);
-
-  useEffect(() => {
-    messagesRef.current = messages;
-  }, [messages]);
-
-  // Snapshot the in-flight chat session into the store on unmount.
-  //
-  // The foreground panel keeps streaming tokens in local React state / refs
-  // (piStreamingTextRef, piContentBlocksRef, messages) — NOT live-mirrored to
-  // the chat-store, for render perf. When the panel unmounts mid-stream because
-  // the user navigated into the standalone /settings route (which tears down
-  // the whole home page), those local tokens would be lost and the background
-  // pi-event router — which takes over once foreground unregisters — would
-  // resume accumulating from a stale store point, leaving a gap in the reply.
-  //
-  // Mirrors the snapshot-on-switch in `loadConversation`: persist the current
-  // messages + streaming cursor so the router continues seamlessly and the
-  // return path (`loadConversation`) rehydrates the full content. Refs (not the
-  // closure values) so the `[]`-deps cleanup reads the latest state. Skipped for
-  // pipe-watch sessions, which are owned by `pipe-watch-writer` (snapshotting
-  // the panel's mirrored copy back would be a lossy round-trip).
-  useEffect(() => {
-    return () => {
-      const sid = piSessionIdRef.current;
-      if (!sid) return;
-      if (!isStreamingRef.current && !isLoadingRef.current) return;
-      const store = useChatStore.getState();
-      const existing = store.sessions[sid];
-      if (!existing || existing.kind === "pipe-watch") return;
-      store.actions.snapshotSession(sid, {
-        messages: messagesRef.current as any,
-        streamingText: piStreamingTextRef.current,
-        streamingMessageId: piMessageIdRef.current,
-        contentBlocks: [...piContentBlocksRef.current],
-        isStreaming: isStreamingRef.current,
-        isLoading: isLoadingRef.current,
-      });
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  // Keep the pipe-context banner in sync with the current session.
-  // When the panel switches AWAY from a pipe-watch session (user
-  // clicks a chat), `activePipeExecution` would otherwise stay set
-  // and the banner would render on top of the chat. Reading the
-  // current session record's kind / pipeContext gives us a single
-  // source of truth tied to conversationId.
-  const currentSessionKind = useChatStore((s) =>
-    s.currentId ? s.sessions[s.currentId]?.kind : undefined,
-  );
-  const currentSessionPipeContext = useChatStore((s) =>
-    s.currentId ? s.sessions[s.currentId]?.pipeContext : undefined,
-  );
-  const { activePipeExecution, startPipeExecution, clearPipeExecution } = useChatPipeWatch({
+  const {
+    activePipeExecution,
+    clearPipeExecution,
     currentSessionKind,
-    currentSessionPipeContext,
+    currentStreamingMessageId,
+    mountedRef,
+    startPipeExecution,
+  } = useChatSessionRuntime({
+    conversationId,
+    setMessages,
+    setIsLoading,
+    setIsStreaming,
+    isLoading,
+    isStreaming,
+    messages,
+    piSessionIdRef,
+    piStreamingTextRef,
+    piMessageIdRef,
+    piContentBlocksRef,
+    isLoadingRef,
+    isStreamingRef,
+    messagesRef,
+    handleAgentEventDataRef,
+    startNewConversationRef,
   });
-
-  // If the Pi pool evicted the session we're currently viewing, swap the
-  // panel to a fresh one. The pool only evicts idle sessions (see
-  // pi.rs::pi_start_inner), so this is rare — but when it does happen the
-  // user shouldn't be left with a panel pointing at a dead pid.
-  useEffect(() => {
-    let cancelled = false;
-    let off: (() => void) | null = null;
-    (async () => {
-      await mountAgentEventBus();
-      if (cancelled) return;
-      off = onAgentEvicted(async (payload) => {
-        if (cancelled) return;
-        if (payload.sessionId !== piSessionIdRef.current) return;
-        await startNewConversationRef.current();
-        emit("chat-current-session", { id: piSessionIdRef.current });
-      });
-    })();
-    return () => {
-      cancelled = true;
-      try { off?.(); } catch { /* ignore */ }
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
 
   const {
     buildProviderConfig,
@@ -1110,2661 +733,193 @@ export function StandaloneChat({
     piPresetSwitchPromiseRef,
   });
 
-  // Focus input on mount
-  useEffect(() => {
-    setTimeout(() => inputRef.current?.focus(), 100);
-  }, []);
-
-  // Escape key: abort agent if running, otherwise close window
-  useEffect(() => {
-    const handleEscape = async (e: KeyboardEvent) => {
-      if (e.key === "Escape" && !showMentionDropdown) {
-        if (isLoading || isStreaming) {
-          // Stop the agent
-          piActiveStopRequestedRef.current = true;
-          try {
-            await commands.piAbortActive(piSessionIdRef.current);
-          } catch (err) {
-            console.warn("[Pi] Failed to abort on Escape:", err);
-          }
-          setIsLoading(false);
-          setIsStreaming(false);
-        } else {
-          commands.closeWindow("Chat");
-        }
-      }
-    };
-    window.addEventListener("keydown", handleEscape);
-    return () => window.removeEventListener("keydown", handleEscape);
-  }, [showMentionDropdown, isLoading, isStreaming]);
-
-  // Apps/tags load on mount, but the first fetch often races server startup.
-  // App names are stable enough to retry only when empty; tags can change
-  // from Brain/timeline while chat is open, so refresh them on menu open.
-  useEffect(() => {
-    if (!appFilterOpen) return;
-    if (appItems.length === 0 && !appsLoading) {
-      void refreshAppItems();
-    }
-    if (!tagsLoading) {
-      void refreshTagItems();
-    }
-  }, [appFilterOpen, appItems.length, appsLoading, tagsLoading, refreshAppItems, refreshTagItems]);
-
-  // Listen for Pi / pipe events.
-  //
-  // Stage 3 of the events refactor: the panel registers with the
-  // agent-event bus instead of subscribing to legacy Tauri topics
-  // directly. Foreground registration is exclusive — the bus routes
-  // events for the registered sessionId to this handler and skips the
-  // background router. See `lib/events/bus.ts`.
-  //
-  // The panel may hold up to two foreground registrations at once:
-  //   - one for the chat session (`conversationId`), bound below in a
-  //     dedicated useEffect that re-registers on every session switch
-  //   - one for a synthetic pipe id (`pipe:<name>:<execId>`) when the
-  //     user is actively watching a pipe — bound inside `initWatch`
-  //     and released on watch end
-  //
-  // The shared `handleAgentEventDataRef` lets both registrations
-  // dispatch through the same event-handling switch without forcing a
-  // costly re-extraction every time the closure changes.
-  useEffect(() => {
-    let unlistenLog: UnlistenFn | null = null;
-    let unlistenReauth: UnlistenFn | null = null;
-    let mounted = true;
-    // Bus registrations to release on cleanup. Mixed with the legacy
-    // unlisten handles below so the cleanup section drains them
-    // uniformly.
-    const busUnregistrations: Array<() => void> = [];
-
-    // Shared handler for Pi event data — used by both pi_event and pipe_event.
-    //
-    // When the rust queue drains a queued prompt, Pi emits text_delta /
-    // thinking_start / tool_execution_start for a NEW turn — but the previous
-    // turn's `agent_end` cleared `piMessageIdRef`, so deltas have no target.
-    // `ensureAssistantPlaceholder` lazily creates one when the trailing message
-    // is a user prompt waiting for a reply (the shape `enqueuePiMessage`
-    // leaves the array in). Idempotent — does nothing when a placeholder
-    // already exists.
-    const ensureAssistantPlaceholder = (): boolean => {
-      if (piMessageIdRef.current) return true;
-      const newAssistantId = (Date.now() + 1).toString();
-      let created = false;
-      setMessages((prev) => {
-        let targetIdx = -1;
-        for (let i = prev.length - 1; i >= 0; i--) {
-          if (prev[i]?.role === "user") {
-            targetIdx = i;
-            break;
-          }
-        }
-        if (targetIdx === -1) return prev;
-
-        const target = prev[targetIdx];
-        if (!target || target.role !== "user") return prev;
-        created = true;
-
-        const base = [...prev];
-        base.splice(targetIdx + 1, 0, {
-          id: newAssistantId,
-          role: "assistant",
-          content: "Processing...",
-          timestamp: Date.now(),
-          model: activePreset?.model,
-          provider: activePreset?.provider,
-        });
-        return base;
-      });
-      if (!created) return false;
-      piMessageIdRef.current = newAssistantId;
-      piStreamingTextRef.current = "";
-      piContentBlocksRef.current = [];
-      setIsLoading(true);
-      setIsStreaming(true);
-      const sidNow = piSessionIdRef.current;
-      if (sidNow) {
-        const storeState = useChatStore.getState();
-        storeState.actions.appendMessage(sidNow, {
-          id: newAssistantId,
-          role: "assistant",
-          content: "Processing...",
-          timestamp: Date.now(),
-          model: activePreset?.model,
-          provider: activePreset?.provider,
-        } as any);
-        storeState.actions.setStreaming(sidNow, {
-          streamingMessageId: newAssistantId,
-          streamingText: "",
-          contentBlocks: [],
-          isStreaming: true,
-          isLoading: true,
-        });
-      }
-      return true;
-    };
-
-    const handlePiEventData = (data: any) => {
-        const emitSessionActivity = (
-          partial: {
-            status?: ReturnType<typeof statusForEvent>;
-            preview?: string;
-            title?: string;
-            lastError?: string;
-            unreadHint?: boolean;
-          },
-          opts?: { throttleMs?: number },
-        ) => {
-          try {
-            const sid = piSessionIdRef.current;
-            if (!sid) return;
-            const status = partial.status ?? null;
-            const preview = partial.preview?.replace(/\s+/g, " ").trim();
-            const title = partial.title?.trim();
-            const lastError = partial.lastError;
-            const unreadHint = partial.unreadHint === true;
-            const updatedAt = Date.now();
-            const sig = `${status ?? ""}|${preview ?? ""}|${title ?? ""}|${lastError ?? ""}|${unreadHint ? "1" : "0"}`;
-            const lastSig = sessionActivityLastSigRef.current[sid];
-            const throttleMs = opts?.throttleMs ?? 0;
-            const lastAt = sessionActivityLastEmitAtRef.current[sid] ?? 0;
-            if (sig === lastSig && throttleMs > 0 && updatedAt - lastAt < throttleMs) return;
-            if (sig === lastSig && throttleMs === 0) return;
-            if (throttleMs > 0 && updatedAt - lastAt < throttleMs && !status && !lastError) return;
-            sessionActivityLastSigRef.current[sid] = sig;
-            sessionActivityLastEmitAtRef.current[sid] = updatedAt;
-            void emit("chat-session-activity", {
-              id: sid,
-              status: status ?? undefined,
-              preview: preview || undefined,
-              title: title || undefined,
-              updatedAt,
-              lastError,
-              unreadHint,
-            });
-          } catch {
-            // best effort only
-          }
-        };
-
-        // Mirror status into the chat-store so the sidebar dot reflects what
-        // Pi is actually doing. The bus routes foreground events exclusively
-        // to this listener, so the router (default subscriber) never sees
-        // them — without this mirror, session.status stays at whatever was
-        // set last time the chat was on the background path (usually
-        // "idle"), and the sidebar shows a hollow gray ring while the
-        // chat is mid-stream. Mirroring covers all of: streaming /
-        // thinking / tool / idle / error per `statusForEvent`.
-        try {
-          const sid = piSessionIdRef.current;
-          const next = sid ? statusForEvent(data) : null;
-          if (sid && next) {
-            const store = useChatStore.getState();
-            const cur = store.sessions[sid]?.status;
-            if (cur !== next) {
-              store.actions.patch(sid, { status: next });
-            }
-            emitSessionActivity({
-              status: next,
-              title: useChatStore.getState().sessions[sid]?.title,
-              lastError: next === "error" ? (piLastErrorRef.current ?? undefined) : undefined,
-            });
-          }
-        } catch {
-          /* defensive — never let a status-mirror failure break the
-             foreground event handler */
-        }
-
-        if (
-          data.type === "agent_end" ||
-          data.type === "pipe_done" ||
-          (data.type === "response" && data.success === false) ||
-          (data.type === "auto_retry_end" && data.success === false) ||
-          (data.type === "message_update" && data.assistantMessageEvent?.type === "error") ||
-          ((data.type === "message_start" || data.type === "message_end") &&
-            data.message?.role === "assistant" &&
-            data.message?.stopReason === "error")
-        ) {
-          cancelStreamingMessageRender();
-        }
-
-        if (
-          data.type === "message_update" &&
-          data.assistantMessageEvent &&
-          data.assistantMessageEvent.type !== "error"
-        ) {
-          const evt = data.assistantMessageEvent;
-          if (evt.type === "text_delta" && evt.delta) {
-            // First delta of a queued turn → create the placeholder lazily.
-            if (!ensureAssistantPlaceholder()) return;
-            piStreamingTextRef.current += evt.delta;
-            emitSessionActivity({
-              status: "streaming",
-              preview: evt.delta,
-              unreadHint: true,
-            }, { throttleMs: 250 });
-
-            // Append to last text block or create new one
-            const blocks = piContentBlocksRef.current;
-            const lastBlock = blocks[blocks.length - 1];
-            if (lastBlock && lastBlock.type === "text") {
-              lastBlock.text += evt.delta;
-            } else {
-              blocks.push({ type: "text", text: evt.delta });
-            }
-
-            scheduleStreamingMessageRender();
-
-          } else if (evt.type === "thinking_start") {
-            if (!ensureAssistantPlaceholder()) return;
-            piThinkingStartRef.current = Date.now();
-            const blocks = piContentBlocksRef.current;
-            blocks.push({ type: "thinking", text: "", isThinking: true });
-            if (piMessageIdRef.current) {
-              const msgId = piMessageIdRef.current;
-              const contentBlocks = [...blocks];
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, content: m.content === "Processing..." ? "" : m.content, contentBlocks } : m)
-              );
-            }
-          } else if (evt.type === "thinking_delta" && evt.delta) {
-            const blocks = piContentBlocksRef.current;
-            const thinkingBlock = blocks[blocks.length - 1];
-            if (thinkingBlock && thinkingBlock.type === "thinking") {
-              thinkingBlock.text += evt.delta;
-            }
-            scheduleStreamingMessageRender();
-          } else if (evt.type === "thinking_end") {
-            const blocks = piContentBlocksRef.current;
-            const thinkingBlock = blocks[blocks.length - 1];
-            if (thinkingBlock && thinkingBlock.type === "thinking") {
-              thinkingBlock.isThinking = false;
-              thinkingBlock.durationMs = piThinkingStartRef.current ? Date.now() - piThinkingStartRef.current : undefined;
-              if (evt.content) thinkingBlock.text = evt.content;
-            }
-            piThinkingStartRef.current = null;
-            if (piMessageIdRef.current) {
-              const msgId = piMessageIdRef.current;
-              const contentBlocks = [...blocks];
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, contentBlocks } : m)
-              );
-            }
-          }
-        } else if (data.type === "tool_execution_start") {
-          if (!ensureAssistantPlaceholder()) return;
-          if (piMessageIdRef.current) {
-            const msgId = piMessageIdRef.current;
-            const toolCall: ToolCall = {
-              id: data.toolCallId || Date.now().toString(),
-              toolName: data.toolName || "unknown",
-              args: data.args || {},
-              isRunning: true,
-            };
-            // Add tool block (text before it is already its own block)
-            piContentBlocksRef.current.push({ type: "tool", toolCall });
-            const contentBlocks = [...piContentBlocksRef.current];
-            setMessages((prev) =>
-              prev.map((m) => m.id === msgId ? { ...m, contentBlocks } : m)
-            );
-          }
-        } else if (data.type === "tool_execution_end") {
-          if (piMessageIdRef.current) {
-            const msgId = piMessageIdRef.current;
-            const toolCallId = data.toolCallId;
-            const resultText = data.result?.content
-              ?.map((c: any) => c.text || "")
-              .join("\n") || "";
-            const truncated = resultText.length > 2000
-              ? resultText.slice(0, 2000) + "\n... (truncated)"
-              : resultText;
-            // Update the tool block in-place
-            for (const block of piContentBlocksRef.current) {
-              if (block.type === "tool" && block.toolCall.id === toolCallId) {
-                block.toolCall.isRunning = false;
-                block.toolCall.result = truncated;
-                block.toolCall.isError = data.isError;
-              }
-            }
-            const contentBlocks = [...piContentBlocksRef.current];
-            setMessages((prev) =>
-              prev.map((m) => m.id === msgId ? { ...m, contentBlocks } : m)
-            );
-          }
-        } else if (data.type === "auto_retry_end" && data.success === false) {
-          // Pi exhausted retries on a transient error (rate limit, overloaded, etc.)
-          const errorStr = data.finalError || "Request failed after retries";
-          console.error("[Pi] Auto-retry failed:", errorStr);
-          piLastErrorRef.current = errorStr;
-          emitSessionActivity({ status: "error", lastError: errorStr });
-
-          // Detect rate limit or daily limit from the error
-          const quotaErrorType = classifyQuotaError(errorStr);
-          if (quotaErrorType === "daily" || quotaErrorType === "rate") {
-            if (quotaErrorType === "daily") {
-              posthog.capture("wall_hit", { reason: "daily_limit", source: "chat" });
-            }
-
-            if (piMessageIdRef.current) {
-              const msgId = piMessageIdRef.current;
-              const content = quotaErrorType === "daily"
-                ? buildDailyLimitMessage(errorStr)
-                : buildRateLimitMessage(errorStr);
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, content } : m)
-              );
-            }
-            // Don't clear refs yet — agent_end will follow
-          } else if (errorStr.includes("model_not_allowed")) {
-            if (piMessageIdRef.current) {
-              const msgId = piMessageIdRef.current;
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, content: "This model requires an upgrade." } : m)
-              );
-            }
-          } else {
-            const providerError = buildProviderErrorMessage(errorStr, activePreset);
-            if (providerError && piMessageIdRef.current) {
-              const msgId = piMessageIdRef.current;
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId
-                  ? { ...m, content: providerError, retryPrompt: lastUserMessageRef.current || undefined }
-                  : m)
-              );
-            }
-          }
-        } else if (data.type === "message_update" && data.assistantMessageEvent?.type === "error") {
-          // Pi's LLM returned an error (e.g. rate limit, overloaded)
-          const reason = data.assistantMessageEvent.reason || "";
-          const errorDetail = data.assistantMessageEvent.error || "";
-          console.error("[Pi] Message error:", reason, errorDetail);
-          emitSessionActivity({ status: "error", lastError: `${reason} ${errorDetail}`.trim() || undefined });
-
-          if (piMessageIdRef.current) {
-            const msgId = piMessageIdRef.current;
-            const fullError = `${reason} ${errorDetail}`.trim();
-            piLastErrorRef.current = fullError;
-
-            const quotaErrorType = classifyQuotaError(fullError);
-            if (quotaErrorType === "daily" || quotaErrorType === "rate") {
-              if (quotaErrorType === "daily") {
-                try {
-                  const match = fullError.match(/"resets_at":\s*"([^"]+)"/);
-                } catch {}
-                                  setMessages((prev) =>
-                  prev.map((m) => m.id === msgId ? { ...m, content: buildDailyLimitMessage(fullError) } : m)
-                );
-              } else {
-                const content = buildRateLimitMessage(fullError);
-                setMessages((prev) =>
-                  prev.map((m) => m.id === msgId ? { ...m, content } : m)
-                );
-              }
-            } else if (fullError.includes("model_not_allowed")) {
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, content: "This model requires an upgrade." } : m)
-              );
-            } else {
-              const providerError = buildProviderErrorMessage(fullError, activePreset);
-              if (providerError) {
-                setMessages((prev) =>
-                  prev.map((m) => m.id === msgId
-                    ? { ...m, content: providerError, retryPrompt: lastUserMessageRef.current || undefined }
-                    : m)
-                );
-              } else if (fullError.includes("already processing")) {
-                // Transient error — Pi was still busy when the prompt arrived.
-                // Don't show it; Pi will process the message once it's free.
-                console.warn("[Pi] Agent busy, waiting for it to finish:", fullError);
-              } else {
-                setMessages((prev) =>
-                  prev.map((m) => m.id === msgId ? { ...m, content: `Error: ${fullError || "Something went wrong"}` } : m)
-                );
-              }
-            }
-          }
-        } else if (data.type === "message_start" && data.message?.role === "user") {
-          // Pi fires `message_start` for each user turn. When a queued
-          // follow-up starts, close the previous streaming target here so the
-          // next text_delta creates a fresh assistant bubble instead of
-          // appending to the prior reply.
-          //
-          // Clear the streaming refs here so the next text_delta lazily
-          // creates a fresh assistant placeholder via `ensureAssistantPlaceholder`.
-          // Skip the very first `message_start (user)` of a run — at that
-          // point `sendPiMessage` has just created an empty placeholder and
-          // there's nothing streamed yet (clearing would orphan the
-          // placeholder and re-create a duplicate on the first delta).
-          const hasStreamedContent =
-            piStreamingTextRef.current.length > 0 ||
-            piContentBlocksRef.current.length > 0;
-          if (hasStreamedContent) {
-            flushStreamingMessageRender();
-            piStreamingTextRef.current = "";
-            piMessageIdRef.current = null;
-            piContentBlocksRef.current = [];
-            // Don't touch isLoading/isStreaming — pi-mono is still busy
-            // processing the followUp turn.
-          }
-
-          const rawText = (() => {
-            const c = data.message?.content;
-            if (typeof c === "string") return c;
-            if (Array.isArray(c)) {
-              return c
-                .filter((p: any) => p?.type === "text" && typeof p.text === "string")
-                .map((p: any) => p.text)
-                .join("");
-            }
-            return "";
-          })();
-          const text = extractConversationHistorySyncUserText(rawText) ?? rawText;
-          const eventImages = imageDataUrlsFromPiContent(data.message?.content);
-          const pendingOptimisticSteer = optimisticSteerRef.current;
-          const isPendingOptimisticSteerEcho = Boolean(
-            pendingOptimisticSteer &&
-            pendingOptimisticSteer.content.trim() === text.trim()
-          );
-          const shouldConsumePendingOptimisticSteer = isPendingOptimisticSteerEcho;
-          const preMatchedTurnIntent = findTurnIntentForUserStart(piSessionIdRef.current, text, pendingNextPiUserDisplayRef.current);
-
-          if (!piMessageIdRef.current || isPendingOptimisticSteerEcho || preMatchedTurnIntent?.kind === "steer") {
-            const sidForStartedUser = piSessionIdRef.current;
-            const pendingDisplay = pendingNextPiUserDisplayRef.current &&
-              (!text || turnIntentTextValuesMatch(pendingNextPiUserDisplayRef.current.preview, text))
-                ? pendingNextPiUserDisplayRef.current
-                : null;
-            const queuedDisplay = pendingDisplay ?? consumeQueuedDisplayForStartedMessage(sidForStartedUser, text);
-            const matchedTurnIntent = preMatchedTurnIntent ?? findTurnIntentForUserStart(sidForStartedUser, text, queuedDisplay);
-            if (matchedTurnIntent?.consumedAssistantId) {
-              pendingNextPiUserIntentRef.current = null;
-              if (pendingNextPiUserDisplayRef.current?.turnIntentId === matchedTurnIntent.id) {
-                pendingNextPiUserDisplayRef.current = null;
-              }
-              if (optimisticSteerRef.current?.turnIntentId === matchedTurnIntent.id) {
-                optimisticSteerRef.current = null;
-              }
-              return;
-            }
-            const queuedImages = queuedDisplay?.images.length ? queuedDisplay.images : eventImages;
-            if (pendingDisplay) {
-              pendingNextPiUserDisplayRef.current = null;
-            }
-            if (!text && !queuedImages.length && !queuedDisplay?.displayContent) {
-              return;
-            }
-            const nextUserIntent = matchedTurnIntent
-              ? (matchedTurnIntent.kind === "steer" ? "steer" : null)
-              : pendingNextPiUserIntentRef.current;
-            pendingNextPiUserIntentRef.current = null;
-            const queuedTurnUserId = Date.now().toString();
-            const queuedTurnAssistantId = (Date.now() + 1).toString();
-            const optimisticSteer = optimisticSteerRef.current;
-            const isOptimisticSteerEcho = Boolean(
-              matchedTurnIntent?.kind === "steer" && matchedTurnIntent.displayedUserId ||
-              queuedDisplay?.optimisticUserId ||
-              (
-                optimisticSteer &&
-                optimisticSteer.content.trim() === text.trim()
-              ),
-            );
-            if (isOptimisticSteerEcho || shouldConsumePendingOptimisticSteer) {
-              optimisticSteerRef.current = null;
-            }
-            if (matchedTurnIntent?.kind === "steer") {
-              markTurnIntentConsumed(matchedTurnIntent.id, queuedTurnAssistantId);
-            }
-            const startedUser: Message | null = isOptimisticSteerEcho ? null : {
-              id: queuedTurnUserId,
-              role: "user",
-              content: text,
-              ...(queuedDisplay?.displayContent ? { displayContent: queuedDisplay.displayContent } : {}),
-              ...(queuedImages.length ? { images: [...queuedImages] } : {}),
-              ...(queuedDisplay?.attachments?.length ? { attachments: [...queuedDisplay.attachments] } : {}),
-              ...(nextUserIntent === "steer" ? { intent: "steer" as const } : {}),
-              ...(matchedTurnIntent ? { turnIntentId: matchedTurnIntent.id } : {}),
-              timestamp: Date.now(),
-            };
-            const assistantPlaceholder: Message = {
-              id: queuedTurnAssistantId,
-              role: "assistant",
-              content: "Processing...",
-              ...(nextUserIntent === "steer" ? { intent: "steer" as const } : {}),
-              ...(matchedTurnIntent ? { turnIntentId: matchedTurnIntent.id } : {}),
-              ...(nextUserIntent === "steer" ? { steeredResponse: true } : {}),
-              timestamp: Date.now(),
-              model: activePreset?.model,
-              provider: activePreset?.provider,
-            };
-
-            let nextRows: Message[] | null = null;
-            setMessages((prev) => {
-              const rows = startedUser
-                ? [...prev, startedUser, assistantPlaceholder]
-                : [...prev, assistantPlaceholder];
-              nextRows = rows;
-              return rows;
-            });
-            if (nextRows) {
-              void saveConversation(nextRows, {
-                refreshHistory: false,
-                syncActiveConversation: false,
-              });
-            }
-
-            piMessageIdRef.current = queuedTurnAssistantId;
-            piStreamingTextRef.current = "";
-            piContentBlocksRef.current = [];
-            setIsLoading(true);
-            setIsStreaming(true);
-
-            if (sidForStartedUser) {
-              const storeState = useChatStore.getState();
-              if (startedUser) {
-                storeState.actions.appendMessage(sidForStartedUser, startedUser as any);
-              }
-              storeState.actions.appendMessage(sidForStartedUser, assistantPlaceholder as any);
-              storeState.actions.setStreaming(sidForStartedUser, {
-                streamingMessageId: queuedTurnAssistantId,
-                streamingText: "",
-                contentBlocks: [],
-                isStreaming: true,
-                isLoading: true,
-              });
-            }
-          }
-        } else if ((data.type === "message_start" || data.type === "message_end") &&
-                   data.message?.role === "assistant" && data.message?.stopReason === "error") {
-          // LLM returned an error (credits_exhausted, rate limit, provider error, etc.)
-          const errMsg = data.message.errorMessage || data.message.error || "Unknown error";
-          console.error("[Pi] LLM error via", data.type, ":", errMsg);
-          piLastErrorRef.current = errMsg;
-          emitSessionActivity({ status: "error", lastError: errMsg });
-          const authTokenInvalidated = isInvalidatedAuthTokenError(errMsg);
-          if (authTokenInvalidated) {
-            void handleInvalidatedAuthToken();
-          }
-
-          if (piMessageIdRef.current) {
-            const msgId = piMessageIdRef.current;
-
-            const quotaErrorType = classifyQuotaError(errMsg);
-            const providerError = buildProviderErrorMessage(errMsg, activePreset);
-            if (authTokenInvalidated) {
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, content: buildInvalidatedAuthTokenMessage() } : m)
-              );
-            } else if (quotaErrorType === "daily") {
-              try {
-                const resetsAtMatch = errMsg.match(/"resets_at":\s*"([^"]+)"/);
-                } catch {}
-                            posthog.capture("wall_hit", { reason: "daily_limit", source: "chat" });
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, content: buildDailyLimitMessage(errMsg) } : m)
-              );
-            } else if (quotaErrorType === "rate") {
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, content: buildRateLimitMessage(errMsg) } : m)
-              );
-            } else if (providerError) {
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId
-                  ? { ...m, content: providerError, retryPrompt: lastUserMessageRef.current || undefined }
-                  : m)
-              );
-            } else {
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, content: `Error: ${errMsg}` } : m)
-              );
-            }
-            piStreamingTextRef.current = "";
-            piMessageIdRef.current = null;
-            piContentBlocksRef.current = [];
-            piLastErrorRef.current = null;
-            setIsLoading(false);
-            setIsStreaming(false);
-          }
-        } else if (data.type === "agent_end") {
-          // When watching a pipe, agent_end fires before pipe_done — don't
-          // clear pipe refs here, let pipe_done handle cleanup instead.
-          const isPipeWatch = piMessageIdRef.current?.startsWith("pipe-");
-          
-          // Always clear loading/streaming state on agent_end, even if piMessageIdRef is null
-          // This fixes the "stuck loading" bug when the ref was cleared prematurely
-          if (!isPipeWatch) {
-            setIsLoading(false);
-            setIsStreaming(false);
-            emitSessionActivity({ status: "idle" });
-          }
-          
-          if (piMessageIdRef.current && !isPipeWatch) {
-            const msgId = piMessageIdRef.current;
-            // Use streamed text if available, otherwise extract from agent_end messages
-            let content = piStreamingTextRef.current;
-
-            // Check for LLM errors in agent_end messages (e.g. credits_exhausted mid-conversation)
-            let agentEndError: string | null = null;
-            if (data.messages && Array.isArray(data.messages)) {
-              for (const m of data.messages) {
-                if (m.role === "assistant" && m.stopReason === "error") {
-                  agentEndError = m.errorMessage || m.error || null;
-                  break;
-                }
-              }
-              if (!content) {
-                // Extract text from all assistant messages in the agent_end payload
-                content = data.messages
-                  .filter((m: any) => m.role === "assistant")
-                  .flatMap((m: any) => (m.content || [])
-                    .filter((c: any) => c.type === "text")
-                    .map((c: any) => c.text))
-                  .join("\n\n");
-              }
-            }
-
-            // Surface credits_exhausted / rate limit / connection errors from agent_end
-            if (agentEndError && !content) {
-              const errStr = agentEndError;
-              const quotaErrorType = classifyQuotaError(errStr);
-              if (isInvalidatedAuthTokenError(errStr)) {
-                void handleInvalidatedAuthToken();
-                content = buildInvalidatedAuthTokenMessage();
-              } else if (quotaErrorType === "daily") {
-                try {
-                  const resetsAtMatch = errStr.match(/"resets_at":\s*"([^"]+)"/);
-                    } catch {}
-                                  content = buildDailyLimitMessage(errStr);
-              } else if (quotaErrorType === "rate") {
-                content = buildRateLimitMessage(errStr);
-              } else if (errStr.includes("model_not_allowed")) {
-                content = "This model requires an upgrade.";
-              } else {
-                content = buildProviderErrorMessage(errStr, activePreset) || errStr;
-              }
-            }
-
-            // Snapshot refs BEFORE setMessages — React's batching may defer the
-            // functional updater until after the refs are cleared below.
-            const blocksSnapshot = [...piContentBlocksRef.current];
-            const streamedText = piStreamingTextRef.current;
-            const wasStoppedByUser = piActiveStopRequestedRef.current;
-
-            // Check if content was already set by error handlers above
-            setMessages((prev) => {
-              const existing = prev.find((m) => m.id === msgId);
-              // Don't overwrite error messages with "Done" or empty content
-              const isErrorMessage = existing?.content?.includes("daily") && existing?.content?.includes("limit") ||
-                existing?.content?.includes("free queries") ||
-                existing?.content?.includes("daily Pro limit") ||
-                existing?.content?.includes("daily query limit") ||
-                existing?.content?.includes("requires an upgrade") ||
-                existing?.content?.includes("Rate limited") ||
-                existing?.content?.includes("rate limit") ||
-                existing?.content?.startsWith("Error:");
-              if (isErrorMessage) {
-                return prev;
-              }
-              // Don't overwrite if we have no new content and existing isn't "Processing..."
-              if (existing && existing.content !== "Processing..." && !content) {
-                return prev;
-              }
-              const contentBlocks = [...blocksSnapshot];
-              if (wasStoppedByUser && !content && contentBlocks.length === 0) {
-                return prev.filter((m) => m.id !== msgId);
-              }
-              // If no text content but we have tool/thinking blocks, don't show "no response"
-              const hasNonTextBlocks = contentBlocks.some((b) => b.type === "tool" || b.type === "thinking");
-              let emptyResponseRetryPrompt: string | undefined;
-              if (!content && hasNonTextBlocks) {
-                content = ""; // empty — tool/thinking blocks will render
-              } else if (!content) {
-                // If any error text was observed during this stream (e.g. a 429
-                // credits_exhausted or daily_cost_limit_exceeded emitted as a
-                // message_update error or auto-retry failure) classify it
-                // before falling back to the generic "no response" string.
-                const lastErr = piLastErrorRef.current;
-                const lastErrKind = lastErr ? classifyQuotaError(lastErr) : "none";
-                if (lastErr && lastErrKind === "daily") {
-                  posthog.capture("wall_hit", { reason: "daily_limit", source: "chat" });
-                  content = buildDailyLimitMessage(lastErr);
-                } else if (lastErr && lastErrKind === "rate") {
-                  content = buildRateLimitMessage(lastErr);
-                } else if (lastErr) {
-                  content = buildProviderErrorMessage(lastErr, activePreset) || `Error: ${lastErr}`;
-                  emptyResponseRetryPrompt = lastUserMessageRef.current || undefined;
-                } else {
-                  content = buildNoResponseMessage(activePreset);
-                  emptyResponseRetryPrompt = lastUserMessageRef.current || undefined;
-                }
-              }
-              // Add text as a content block if no text block exists yet
-              const hasTextBlock = contentBlocks.some((b) => b.type === "text");
-              if (!streamedText && content && !hasTextBlock) {
-                contentBlocks.push({ type: "text", text: content });
-              }
-              return prev.map((m) => m.id === msgId
-                ? { ...m, content, contentBlocks, ...(emptyResponseRetryPrompt ? { retryPrompt: emptyResponseRetryPrompt } : {}) }
-                : m);
-            });
-            if (!isPipeWatch) {
-              const analyticsPayload = {
-                provider: activePreset?.provider,
-                model: activePreset?.model,
-                has_tool_use: blocksSnapshot.some((b) => b.type === "tool"),
-                response_length: streamedText?.length ?? 0,
-              };
-              setTimeout(() => {
-                posthog.capture("chat_response_received", analyticsPayload);
-              }, POST_STREAM_SIDE_EFFECT_DELAY_MS);
-            }
-          }
-          if (!isPipeWatch) {
-            piStreamingTextRef.current = "";
-            piMessageIdRef.current = null;
-            piContentBlocksRef.current = [];
-            piLastErrorRef.current = null;
-            piActiveStopRequestedRef.current = false;
-            piThinkingStartRef.current = null;
-            forceQueueModeRef.current = false;
-            piRateLimitRetries.current = 0;
-            setIsLoading(false);
-            setIsStreaming(false);
-            emitSessionActivity({ status: "idle" });
-            if (pendingSteerBatchRef.current.some((item) => item.sessionId === piSessionIdRef.current)) {
-              void flushPendingSteerBatch();
-            }
-          }
-        } else if (data.type === "response" && data.success === false) {
-          const errorStr = data.error || "Unknown error";
-          emitSessionActivity({ status: "error", lastError: errorStr });
-          // Pi agent first-call bug (pi-mono#2461) — first RPC prompt crashes.
-          // Auto-retry the same prompt once. The second call works.
-          if (errorStr.includes("startsWith") || errorStr.includes("text.startsWith")) {
-            console.warn("[Pi] first-call bug hit, auto-retrying prompt:", errorStr);
-            if (piMessageIdRef.current && !piFirstCallRetried.current) {
-              piFirstCallRetried.current = true;
-              // Re-send the last prompt
-              const lastUserMsg = messages.findLast(m => m.role === "user");
-              if (lastUserMsg?.content) {
-                commands.piPrompt(piSessionIdRef.current, lastUserMsg.content, null, null).catch(() => {});
-              }
-            }
-            return;
-          }
-          // Rate-limit (429) auto-retry — honor the gateway's reset_in hint and
-          // re-send the same prompt. The cloud LLM gateway caps free/logged-in
-          // tiers at a few dozen requests/minute; a single agentic run can trip
-          // it, after which a short wait clears the budget. Without this the turn
-          // dies silently (e.g. pipe creation stalls mid-skill).
-          if (
-            classifyQuotaError(errorStr) === "rate" &&
-            piRateLimitRetries.current < PI_MAX_RATE_LIMIT_RETRIES &&
-            piSessionIdRef.current &&
-            lastUserMessageRef.current
-          ) {
-            piRateLimitRetries.current += 1;
-            const attempt = piRateLimitRetries.current;
-            const waitSecs = parseRateLimitWaitSeconds(errorStr);
-            const retrySession = piSessionIdRef.current;
-            const retryPrompt = lastUserMessageRef.current;
-            console.warn(`[Pi] rate limited, auto-retry ${attempt}/${PI_MAX_RATE_LIMIT_RETRIES} in ${waitSecs}s:`, errorStr);
-            // Reset the in-flight buffers so the retried turn renders cleanly into
-            // the same bubble instead of appending onto any pre-429 partial output.
-            piStreamingTextRef.current = "";
-            piContentBlocksRef.current = [];
-            const retryTurnId = piMessageIdRef.current;
-            if (retryTurnId) {
-              setMessages((prev) =>
-                prev.map((m) => m.id === retryTurnId
-                  ? { ...m, content: `Rate limited — retrying in ${waitSecs}s… (attempt ${attempt}/${PI_MAX_RATE_LIMIT_RETRIES})`, contentBlocks: [] }
-                  : m)
-              );
-            }
-            setTimeout(() => {
-              // Guard the delayed re-send: bail if the user unmounted, switched
-              // sessions, or started a new turn during the wait, so we never
-              // inject a stale prompt into the wrong place.
-              if (!mountedRef.current) return;
-              if (piSessionIdRef.current !== retrySession) return;
-              if (piMessageIdRef.current && piMessageIdRef.current !== retryTurnId) return;
-              commands.piPrompt(retrySession, retryPrompt, null, null).catch(() => {});
-            }, waitSecs * 1000);
-            return;
-          }
-          if (piMessageIdRef.current) {
-            const msgId = piMessageIdRef.current;
-
-            const quotaErrorType = classifyQuotaError(errorStr);
-            if (quotaErrorType === "daily" || quotaErrorType === "rate") {
-              if (quotaErrorType === "daily") {
-                try {
-                  const match = errorStr.match(/"resets_at":\s*"([^"]+)"/);
-                } catch {}
-                                  setMessages((prev) =>
-                  prev.map((m) => m.id === msgId ? { ...m, content: buildDailyLimitMessage(errorStr) } : m)
-                );
-              } else {
-                const content = buildRateLimitMessage(errorStr);
-                setMessages((prev) =>
-                  prev.map((m) => m.id === msgId ? { ...m, content } : m)
-                );
-              }
-            } else if (errorStr.includes("model_not_allowed")) {
-              setMessages((prev) =>
-                prev.map((m) => m.id === msgId ? { ...m, content: "This model requires an upgrade." } : m)
-              );
-            } else {
-              const providerError = buildProviderErrorMessage(errorStr, activePreset);
-              if (providerError) {
-                setMessages((prev) =>
-                  prev.map((m) => m.id === msgId
-                    ? { ...m, content: providerError, retryPrompt: lastUserMessageRef.current || undefined }
-                    : m)
-                );
-              } else if (errorStr.includes("already processing")) {
-                console.warn("[Pi] already-processing race in response event:", errorStr);
-                setMessages((prev) =>
-                  prev.map((m) => m.id === msgId ? {
-                    ...m,
-                    content: "The AI was mid-response when your message arrived.",
-                    retryPrompt: lastUserMessageRef.current || undefined,
-                  } : m)
-                );
-              } else if (errorStr.includes("api_error") || errorStr.includes("Internal server error") || /\b5\d\d\b/.test(errorStr)) {
-                // Upstream API 5xx — SDK already exhausted its auto-retry attempts
-                setMessages((prev) =>
-                  prev.map((m) => m.id === msgId ? {
-                    ...m,
-                    content: "Something went wrong on the server.",
-                    retryPrompt: lastUserMessageRef.current || undefined,
-                  } : m)
-                );
-              } else {
-                setMessages((prev) =>
-                  prev.map((m) => m.id === msgId ? { ...m, content: `Error: ${errorStr}` } : m)
-                );
-              }
-            }
-          }
-          const quotaErrorType = classifyQuotaError(errorStr);
-          const errorCategory = quotaErrorType === "daily" ? "daily_limit"
-            : quotaErrorType === "rate" ? "rate_limit"
-            : errorStr.includes("model_not_allowed") ? "model_not_allowed"
-            : "other";
-          posthog.capture("chat_response_error", {
-            provider: activePreset?.provider,
-            model: activePreset?.model,
-            error_type: errorCategory,
-          });
-          piStreamingTextRef.current = "";
-          optimisticSteerRef.current = null;
-          if (piMessageIdRef.current?.startsWith("pipe-")) {
-            clearPipeExecution();
-          }
-          piMessageIdRef.current = null;
-          piContentBlocksRef.current = [];
-          setIsLoading(false);
-          setIsStreaming(false);
-        } else if (data.type === "pipe_done") {
-          // Pipe execution finished — clean up streaming state
-          if (piMessageIdRef.current?.startsWith("pipe-")) {
-            const msgId = piMessageIdRef.current;
-            const content = piStreamingTextRef.current || "Pipe completed with no output.";
-            const blocksSnapshot = [...piContentBlocksRef.current];
-            setMessages((prev) =>
-              prev.map((m) => m.id === msgId ? { ...m, content, contentBlocks: blocksSnapshot } : m)
-            );
-            piStreamingTextRef.current = "";
-            piMessageIdRef.current = null;
-            piContentBlocksRef.current = [];
-            piLastErrorRef.current = null;
-            piThinkingStartRef.current = null;
-            clearPipeExecution();
-            setIsLoading(false);
-            setIsStreaming(false);
-          }
-        }
-      };
-
-    // Publish the current handler to the forwarding ref so foreground
-    // registrations (chat + pipe-watch) dispatch through the same
-    // closure without re-binding.
-    handleAgentEventDataRef.current = handlePiEventData;
-
-    const setup = async () => {
-      // Ensure the bus's Tauri listener is up before any consumer
-      // (router, panel, pipes hook) starts registering. Idempotent.
-      await mountAgentEventBus();
-
-      // Termination — broadcast event, filter by current session id.
-      // Replaces the prior `listen("pi_terminated", ...)`. The bus
-      // mirrors `agent_terminated`; legacy `pi_terminated` is a Stage 5
-      // cleanup target.
-      busUnregistrations.push(onAgentTerminated(async (payload) => {
-        if (!mounted) return;
-        if (payload.sessionId !== piSessionIdRef.current) return;
-        const terminatedPid = payload.pid;
-        const termKey = `${payload.sessionId}:${typeof terminatedPid === "number" ? terminatedPid : "unknown"}`;
-        const nowMs = Date.now();
-        const lastSeen = piTerminationDedupRef.current[termKey] ?? 0;
-        if (nowMs - lastSeen < 4000) {
-          return;
-        }
-        piTerminationDedupRef.current[termKey] = nowMs;
-        if (typeof terminatedPid === "number" && piIntentionallyStoppedPidsRef.current.delete(terminatedPid)) {
-          return;
-        }
-        if (piStoppedIntentionallyRef.current) {
-          piStoppedIntentionallyRef.current = false;
-          return;
-        }
-        console.log("[Pi] Process terminated, pid:", terminatedPid);
-        try {
-          const info = await commands.piInfo(piSessionIdRef.current);
-          if (info.status === "ok" && info.data.running && info.data.pid !== terminatedPid) {
-            setPiInfo(info.data);
-            return;
-          }
-        } catch {}
-
-        // If a message was in flight, append error to the message so the user
-        // knows the agent stopped unexpectedly (not just "completed").
-        if (piMessageIdRef.current) {
-          const msgId = piMessageIdRef.current;
-          setMessages((prev) =>
-            prev.map((m) => {
-              if (m.id !== msgId) return m;
-              const existing = m.content && m.content !== "Processing..." ? m.content : "";
-              const errorSuffix = "\n\n---\n\n⚠️ agent stopped unexpectedly — restarting automatically...";
-              return { ...m, content: existing + errorSuffix };
-            })
-          );
-          piStreamingTextRef.current = "";
-          piMessageIdRef.current = null;
-          piContentBlocksRef.current = [];
-          setIsLoading(false);
-          setIsStreaming(false);
-        }
-
-        // Auto-restart with exponential backoff to avoid crash loops
-        const now = Date.now();
-        const MAX_CRASHES = 5;
-        const CRASH_WINDOW_MS = 60_000; // reset counter after 1 min of stability
-
-        // Reset crash counter if last crash was long ago (Pi was stable)
-        if (now - piLastCrashRef.current > CRASH_WINDOW_MS) {
-          piCrashCountRef.current = 0;
-        }
-        piCrashCountRef.current++;
-        piLastCrashRef.current = now;
-
-        if (piCrashCountRef.current > MAX_CRASHES) {
-          console.error(`[Pi] Crash loop detected (${piCrashCountRef.current} crashes in ${CRASH_WINDOW_MS / 1000}s) — stopping auto-restart. User action required.`);
-          setPiInfo(null);
-          return;
-        }
-
-        // Exponential backoff: 1.5s, 3s, 6s, 12s, 24s
-        const delay = 1500 * Math.pow(2, piCrashCountRef.current - 1);
-        console.log(`[Pi] Auto-restart attempt ${piCrashCountRef.current}/${MAX_CRASHES} in ${delay}ms`);
-
-        setTimeout(async () => {
-          if (!mounted) return;
-          // Check if a newer Pi process is already running (race: stop → start → terminated)
-          try {
-            const result = await commands.piInfo(piSessionIdRef.current);
-            if (result.status === "ok" && result.data.running && result.data.pid !== terminatedPid) {
-              console.log("[Pi] Stale termination for pid", terminatedPid, "— newer pid", result.data.pid, "is running");
-              setPiInfo(result.data);
-              return;
-            }
-          } catch {}
-
-          if (!piStartInFlightRef.current) {
-            console.log("[Pi] Auto-restarting after crash");
-            try {
-              const providerConfig = buildProviderConfig();
-              const home = await homeDir();
-              const dir = await join(home, ".screenpipe", "pi-chat");
-              const result = await commands.piStart(piSessionIdRef.current, dir, settings.user?.token ?? null, providerConfig);
-              if (result.status === "ok") {
-                setPiInfo(result.data);
-                piSessionSyncedRef.current = false;
-                // Keep running-config ref in sync so preset watcher doesn't re-trigger
-                if (providerConfig) {
-                  piRunningConfigRef.current = {
-                    provider: providerConfig.provider,
-                    model: providerConfig.model,
-                    url: providerConfig.url,
-                    apiKey: providerConfig.apiKey,
-                    maxTokens: providerConfig.maxTokens,
-                    systemPrompt: providerConfig.systemPrompt,
-                    token: settings.user?.token ?? null,
-                  };
-                }
-              } else {
-                console.error("[Pi] Auto-restart failed:", result.error);
-                setPiInfo(null);
-              }
-            } catch (e) {
-              console.error("[Pi] Auto-restart exception:", e);
-              setPiInfo(null);
-            }
-          }
-        }, delay);
-      }));
-      // Listen for Pi stderr — only surface errors when user is actively waiting for a response
-      unlistenLog = await listen<string>("pi_log", (event) => {
-        if (!mounted) return;
-        // Only show errors if user sent a message and is waiting — not during background startup/restart
-        if (!piMessageIdRef.current) return;
-        const line = event.payload;
-        if (line.includes("model_not_allowed") || line.includes("403")) {
-          cancelStreamingMessageRender();
-          const msgId = piMessageIdRef.current;
-          if (msgId) {
-            setMessages((prev) =>
-              prev.map((m) => m.id === msgId ? { ...m, content: "This model requires an upgrade — try a different model in your AI preset." } : m)
-            );
-          }
-        } else if (line.includes("429") || line.includes("rate") || line.includes("daily_limit")) {
-          cancelStreamingMessageRender();
-          const msgId = piMessageIdRef.current;
-          if (msgId) {
-            setMessages((prev) =>
-              prev.map((m) => m.id === msgId ? { ...m, content: "Rate limited — try again in a moment or switch to a different model." } : m)
-            );
-          }
-        } else if (line.includes("content must be a string") || line.includes("does not support images") || line.includes("image_url is not supported")) {
-          cancelStreamingMessageRender();
-          const msgId = piMessageIdRef.current;
-          if (msgId) {
-            setMessages((prev) =>
-              prev.map((m) => m.id === msgId ? { ...m, content: "This model doesn't support images — try a vision-capable model (e.g. llama-4-scout on Groq, gpt-4o on OpenAI)." } : m)
-            );
-          }
-        } else if (line.includes("not found") || line.includes("ECONNREFUSED") || line.includes("connection refused")) {
-          let hint = line;
-          if (line.includes("not found")) {
-            hint = `Model not found: ${line}. Check your AI preset in settings.`;
-          } else if (line.includes("ECONNREFUSED") || line.includes("connection refused")) {
-            hint = "Cannot connect to Ollama — is it running? Start with: ollama serve";
-          }
-          toast({ title: "Pi agent error", description: hint, variant: "destructive" });
-        }
-      });
-    };
-
-    setup();
-
-    // Restart the current session when a new auth token arrives (deeplink login).
-    listen<{ apiKey: string }>("pi-reauth", async (event) => {
-      if (!mounted) return;
-      try {
-        const home = await homeDir();
-        const dir = await join(home, ".screenpipe", "pi-chat");
-        const result = await commands.piStart(piSessionIdRef.current, dir, event.payload.apiKey, buildProviderConfig());
-        if (result.status === "ok") {
-          setPiInfo(result.data);
-          piSessionSyncedRef.current = false;
-          syncThinkingLevelAfterStart(piSessionIdRef.current);
-        }
-      } catch (e) {
-        console.warn("[Pi] reauth restart skipped:", e);
-      }
-    }).then(fn => { unlistenReauth = fn; });
-
-    return () => {
-      mounted = false;
-      cancelStreamingMessageRender();
-      for (const off of busUnregistrations) {
-        try { off(); } catch { /* ignore — tearing down */ }
-      }
-      unlistenLog?.();
-      unlistenReauth?.();
-      // Deliberately do NOT abort the Pi session here. Unmount happens when
-      // the user navigates away from chat (e.g. into the standalone /settings
-      // route, which unmounts the whole home page). Aborting would kill an
-      // in-flight response — the exact regression users hit ("opening Settings
-      // stops the current chat"). Instead we let the session keep streaming:
-      //   - the app-lifetime pi-event router (registerDefault) takes over once
-      //     this panel releases its foreground registration and accumulates
-      //     tokens into the chat-store while we're away;
-      //   - on return, `loadConversation` rehydrates that background-streamed
-      //     state and re-registers foreground, resuming exactly where we left.
-      // The old "already processing" hazard this guarded against is now handled
-      // by the Rust command queue (pi_command_queue.rs), which serializes/queues
-      // prompts instead of rejecting them.
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  // Watch pipe: set up tracking from either Tauri event or sessionStorage (for cross-page navigation)
-  useEffect(() => {
-    let watchPollTimer: ReturnType<typeof setTimeout> | null = null;
-
-    // Poll execution API to check if pipe already finished (race condition fix)
-    const pollExecutionStatus = async (pipeName: string, executionId: number, pipeSid: string) => {
-      try {
-        const res = await localFetch(`/pipes/${pipeName}/executions?limit=20`);
-        if (!res.ok) return;
-        const data = await res.json();
-        const exec = (data.data || []).find((e: any) => e.id === executionId);
-        if (!exec) return;
-
-        // Pipe already finished before live events could reach the writer
-        // (race between pipe completion and bus mount). Reconstruct the
-        // conversation from stdout and write it directly to chat-store —
-        // the panel mirrors store messages for pipe-watch sessions, so
-        // this surfaces the result without a separate render path.
-        if (exec.status !== "running") {
-          const { parsePipeNdjsonToMessages } = await import(
-            "@/lib/pipe-ndjson-to-chat"
-          );
-          let messagesFromStdout = exec.stdout
-            ? parsePipeNdjsonToMessages(exec.stdout)
-            : [];
-          if (messagesFromStdout.length === 0) {
-            const fallback =
-              exec.status === "failed"
-                ? `Pipe failed: ${exec.error_message || exec.stderr || "unknown error"}`
-                : "Pipe completed with no output.";
-            messagesFromStdout = [
-              {
-                id: `pipe-poll-${executionId}`,
-                role: "assistant",
-                content: fallback,
-                timestamp: Date.now(),
-              } as any,
-            ];
-          }
-          const store = useChatStore.getState();
-          if (store.sessions[pipeSid]) {
-            store.actions.setMessages(pipeSid, messagesFromStdout as any);
-            store.actions.endTurn(pipeSid);
-          }
-          return true;
-        }
-        return false; // still running
-      } catch {
-        return false;
-      }
-    };
-
-    const initWatch = async (pipeName: string, executionId: number, presetId?: string | null) => {
-      startPipeExecution(pipeName, executionId);
-
-      // Apply the pipe's AI preset so the chat header reflects it
-      if (presetId && settings.aiPresets) {
-        const match = settings.aiPresets.find((p) => p.id === presetId);
-        if (match) setActivePreset(match);
-      }
-
-      const pipeSid = pipeSessionId(pipeName, executionId);
-
-      // Pipe-watch is a real session (kind: "pipe-watch"). The writer
-      // (`pipe-watch-writer`) is the sole authority for its message
-      // content — it implicit-creates messages on first content event
-      // and prefers `agent_end`'s authoritative messages array on
-      // terminal events. We upsert the session record synchronously
-      // here so the writer can identify the sid as kind=pipe-watch
-      // for any events that arrive between this call and
-      // loadConversation finishing its async setup.
-      const startedAt = new Date().toISOString();
-      const storeNow = useChatStore.getState();
-      if (!storeNow.sessions[pipeSid]) {
-        storeNow.actions.upsert({
-          id: pipeSid,
-          title: pipeName,
-          preview: "",
-          status: "streaming",
-          messageCount: 0,
-          createdAt: Date.now(),
-          updatedAt: Date.now(),
-          pinned: false,
-          unread: false,
-          kind: "pipe-watch",
-          pipeContext: { pipeName, executionId, startedAt },
-          isLoading: true,
-          isStreaming: true,
-        });
-      }
-
-      const pipeConv: ChatConversation = {
-        id: pipeSid,
-        title: pipeName,
-        // No placeholder — the writer creates the first message on the
-        // first real content event. Until then the panel shows a
-        // loading indicator (isLoading=true) which matches the visual
-        // we want during pipe startup.
-        messages: [],
-        createdAt: Date.now(),
-        updatedAt: Date.now(),
-        kind: "pipe-watch",
-        pipeContext: { pipeName, executionId, startedAt },
-      };
-      await loadConversationRef.current(pipeConv);
-
-      // No piMessageIdRef setup — the writer owns message lifecycle
-      // for pipe-watch. The local refs stay null/empty so the chat
-      // panel's chat-shaped event handlers (which only fire if
-      // foreground is registered, which it isn't for pipe-watch)
-      // can't accidentally write to a stale placeholder id.
-      setIsStreaming(true);
-      setIsLoading(true);
-
-      // Poll the executions API as a safety net — catches the case
-      // where the pipe finished BEFORE we mounted the foreground bus
-      // registration (the events fired and went to the recorder, not
-      // here). Once the live agent_event stream has had a chance to
-      // arrive, this poll has done its job; the live stream is the
-      // authoritative source for in-progress runs.
-      //
-      // Bug fix (2026-04-26): the previous version tore down the watch
-      // after 30s "timeout" — clearing activePipeExecution, unregistering
-      // the foreground, and nulling piMessageIdRef. For pipes that take
-      // longer than 30s this would silently (a) hide the banner, (b)
-      // strand the thinking indicator at isThinking:true, and (c) drop
-      // every subsequent live event on the floor because piMessageIdRef
-      // was null. Now we just stop polling — the watch stays alive and
-      // is driven by live events to completion.
-      let pollCount = 0;
-      const maxPolls = 10; // 30s of safety-net polling
-      const doPoll = async () => {
-        // Stop polling if the user navigated to a different chat. The
-        // writer still accumulates events for this sid in the
-        // background — we just don't need the poll fallback once we're
-        // not actively viewing.
-        if (piSessionIdRef.current !== pipeSid) return;
-        const done = await pollExecutionStatus(pipeName, executionId, pipeSid);
-        if (done) {
-          watchPollTimer = null;
-          return;
-        }
-        pollCount++;
-        if (pollCount >= maxPolls) {
-          // Safety-net budget exhausted. The pipe is running and live
-          // events are doing their job — no teardown here. The watch
-          // ends when agent_end / pipe_done arrives via the bus.
-          watchPollTimer = null;
-          return;
-        }
-        watchPollTimer = setTimeout(doPoll, 3000);
-      };
-      // Small delay before first poll to let streaming events arrive first
-      watchPollTimer = setTimeout(doPoll, 1500);
-    };
-
-    // Check sessionStorage first (set by pipes-section before navigation)
-    const stored = sessionStorage.getItem("watchPipe");
-    if (stored) {
-      sessionStorage.removeItem("watchPipe");
-      try {
-        const { pipeName, executionId, presetId } = JSON.parse(stored);
-        if (pipeName && executionId != null) {
-          initWatch(pipeName, executionId, presetId);
-        }
-      } catch {}
-    }
-
-    // Also listen for live events (in case chat is already mounted)
-    let unlisten: (() => void) | null = null;
-    listen<{ pipeName: string; executionId: number; presetId?: string | null }>("watch_pipe", (event) => {
-      const { pipeName, executionId, presetId } = event.payload;
-      initWatch(pipeName, executionId, presetId);
-    }).then((fn) => { unlisten = fn; });
-    return () => {
-      unlisten?.();
-      if (watchPollTimer) clearTimeout(watchPollTimer);
-    };
-  }, []);
-
-  // Send message using Pi agent
-  /**
-   * Enqueue a follow-up while another prompt is still streaming.
-   *
-   * Distinct from `sendPiMessage` because we explicitly want to NOT:
-   *   - abort the in-flight turn,
-   *   - clobber `piMessageIdRef` / streaming refs,
-   *   - flip `isLoading`/`isStreaming`.
-   *
-   * The rust queue (`pi_command_queue.rs`) holds the prompt until the current
-   * turn's `agent_end` fires, then writes it to stdin. The pi-event-router
-   * downstream picks up the new turn and appends user/assistant content to
-   * the chat-store — same path used for any other message, just kicked off
-   * after the queue drains.
-   */
-  async function enqueuePiMessage(userMessage: string, displayLabel?: string, imageDataUrls?: string[]) {
-    if (!piInfo?.running) {
-      // No Pi running → fall back to the normal start-and-send path.
-      return sendPiMessage(userMessage, displayLabel, imageDataUrls);
-    }
-
-    // Convert any data-URL pastes to the Pi image-content shape (same format
-    // used by the normal send path further down in this file).
-    const outgoingImages = imageDataUrls ?? pastedImages;
-    const piImages = imageDataUrlsToPiImages(outgoingImages);
-    const queuedImageDataUrls = outgoingImages.length > 0 ? [...outgoingImages] : [];
-    const prevInput = input;
-    const hadPastedImages = imageDataUrls == null && pastedImages.length > 0;
-    // Snapshot whatever sendMessage stashed for us. Consumed here so it
-    // doesn't leak into a later turn if this enqueue races with another.
-    const queuedAttachments = consumePendingAttachments();
-
-    setInput("");
-    if (inputRef.current) inputRef.current.style.height = "auto";
-    if (hadPastedImages) setPastedImages([]);
-
-    // Issue #3636: same contract as sendPiMessage's send path — every
-    // turn carries the recent conversation history so the model has
-    // context even if Pi's internal session lost it (compaction,
-    // crash + auto-restart, kill that the termination handler missed).
-    // The queue path was previously a silent gap: when an earlier send
-    // was still in-flight, follow-ups routed here got the bare user
-    // message, and any Pi state divergence in between manifested as
-    // "chat suddenly forgot what we were talking about."
-    const queuedPrompt = withConversationHistory(userMessage, messages);
-
-    // E2E test hook — capture queued prompts for context-loss assertions
-    {
-      const g = window as any;
-      if (Array.isArray(g.__e2ePiPromptCaptures)) {
-        g.__e2ePiPromptCaptures.push({
-          sessionId: piSessionIdRef.current,
-          message: queuedPrompt,
-          at: Date.now(),
-        });
-      }
-    }
-
-    try {
-      const result = await commands.piQueuePrompt(
-        piSessionIdRef.current,
-        queuedPrompt,
-        piImages.length > 0 ? piImages : null,
-        queuedPreviewForText(userMessage),
-      );
-      const queuedTurnIntentId = `queued-${result.status === "ok" ? result.data : Date.now()}`;
-      if (result.status !== "ok") {
-        setInput(prevInput);
-        if (hadPastedImages) setPastedImages(queuedImageDataUrls);
-        toast({ title: "failed to queue message", description: result.error, variant: "destructive" });
-        return;
-      }
-
-      registerTurnIntent({
-        id: queuedTurnIntentId,
-        sessionId: piSessionIdRef.current,
-        kind: "queued",
-        content: userMessage,
-        preview: queuedPreviewForText(userMessage),
-        queueId: result.data,
-        createdAt: Date.now(),
-      });
-      restoreQueuedDisplay(piSessionIdRef.current, result.data, {
-        preview: queuedPreviewForText(userMessage),
-        images: queuedImageDataUrls,
-        ...(queuedAttachments ? { attachments: queuedAttachments } : {}),
-        ...(displayLabel ? { displayContent: displayLabel } : {}),
-        turnIntentId: queuedTurnIntentId,
-      });
-
-      posthog.capture("chat_message_enqueued", {
-        provider: activePreset?.provider,
-        model: activePreset?.model,
-        pending_count: queuedPrompts.length + 1,
-      });
-    } catch (e) {
-      setInput(prevInput);
-      if (hadPastedImages) setPastedImages(queuedImageDataUrls);
-      console.warn("[Pi] failed to enqueue follow-up:", e);
-    }
-  }
-
-  function clearActivePiTurnState() {
-    cancelStreamingMessageRender();
-    piStreamingTextRef.current = "";
-    piMessageIdRef.current = null;
-    piContentBlocksRef.current = [];
-    forceQueueModeRef.current = false;
-    setIsLoading(false);
-    setIsStreaming(false);
-  }
-
-  async function interruptActivePiTurn() {
-    const hasActiveTurn = isLoading || isStreaming || !!piMessageIdRef.current;
-    if (!hasActiveTurn) return;
-
-    let aborted = false;
-    try {
-      const abortResult = await Promise.race([
-        commands.piAbort(piSessionIdRef.current),
-        new Promise<{ status: "error"; error: string }>((resolve) => {
-          window.setTimeout(() => resolve({ status: "error", error: "Abort timed out" }), 1_500);
-        }),
-      ]);
-      if (abortResult.status === "ok") {
-        aborted = true;
-      } else {
-        console.warn("[Pi] Abort before steering failed:", abortResult.error);
-      }
-    } catch (e) {
-      console.warn("[Pi] Abort before steering threw:", e);
-    }
-
-    if (!aborted) {
-      const providerConfig = buildProviderConfig();
-      if (!providerConfig) {
-        throw new Error("No AI preset selected");
-      }
-      await restartCurrentPiSession(providerConfig);
-    }
-
-    clearActivePiTurnState();
-  }
-
-  async function sendPiMessage(userMessage: string, displayLabel?: string, imageDataUrls?: string[]) {
-    clearPendingSteerTransportState();
-
-    // Auto-start Pi if it's not running yet (new session or crash recovery)
-    if (!piInfo?.running) {
-      if (piStartInFlightRef.current) {
-        if (!autoSendBypassRef.current) {
-          toast({ title: "Pi starting", description: "Please wait a moment", variant: "destructive" });
-          return;
-        }
-        // Prefill auto-send: wait for in-flight start to complete
-        const startWait = Date.now();
-        while (piStartInFlightRef.current && Date.now() - startWait < 10000) {
-          await new Promise(r => setTimeout(r, 300));
-        }
-        if (piStartInFlightRef.current) return; // timed out
-      } else {
-        console.log("[Pi] Not running, auto-starting before sending message");
-        piStartInFlightRef.current = true;
-        setPiStarting(true);
-        const providerConfig = buildProviderConfig();
-        try {
-          const home = await homeDir();
-          const dir = await join(home, ".screenpipe", "pi-chat");
-          const result = await commands.piStart(piSessionIdRef.current, dir, settings.user?.token ?? null, providerConfig);
-          if (result.status === "ok" && result.data.running) {
-            setPiInfo(result.data);
-            piSessionSyncedRef.current = false;
-            piCrashCountRef.current = 0; // reset crash loop counter on manual start
-            // Keep running-config ref in sync so preset watcher doesn't re-trigger
-            if (providerConfig) {
-              setRunningConfigFromProviderConfig(providerConfig);
-            }
-            syncThinkingLevelAfterStart(piSessionIdRef.current);
-          } else {
-            const providerLabel = providerConfig?.provider || "AI";
-            toast({ title: `failed to start AI assistant (${providerLabel})`, description: result.status === "error" ? result.error : "Unknown error", variant: "destructive" });
-            return;
-          }
-        } catch (e) {
-          const providerLabel = providerConfig?.provider || "AI";
-          toast({ title: `failed to start AI assistant (${providerLabel})`, description: String(e), variant: "destructive" });
-          return;
-        } finally {
-          setPiStarting(false);
-          piStartInFlightRef.current = false;
-        }
-      }
-    }
-
-    if (piPresetSwitchPromiseRef.current) {
-      await piPresetSwitchPromiseRef.current;
-    }
-
-    await interruptActivePiTurn();
-    forceQueueModeRef.current = true;
-
-    const outgoingImages = imageDataUrls ?? pastedImages;
-    const shouldClearPastedImages = imageDataUrls == null && pastedImages.length > 0;
-
-    const consumedAttachments = consumePendingAttachments();
-    const newUserMessage: Message = {
-      id: Date.now().toString(),
-      role: "user",
-      content: userMessage,
-      ...(displayLabel ? { displayContent: displayLabel } : {}),
-      ...(outgoingImages.length > 0 ? { images: [...outgoingImages] } : {}),
-      ...(consumedAttachments ? { attachments: consumedAttachments } : {}),
-      timestamp: Date.now(),
-    };
-
-    const assistantMessageId = (Date.now() + 1).toString();
-
-    piStreamingTextRef.current = "";
-    piMessageIdRef.current = assistantMessageId;
-    piContentBlocksRef.current = [];
-
-    piRateLimitRetries.current = 0;
-    lastUserMessageRef.current = userMessage;
-
-    let nextRowsAfterUserAppend: Message[] | null = null;
-    setMessages((prev) => {
-      const next = [...prev, newUserMessage];
-      nextRowsAfterUserAppend = next;
-      return next;
-    });
-    if (nextRowsAfterUserAppend) {
-      void saveConversation(nextRowsAfterUserAppend, { refreshHistory: false });
-    }
-    setInput("");
-    if (inputRef.current) inputRef.current.style.height = "auto";
-    setIsLoading(true);
-    setIsStreaming(true);
-
-    // Mirror the user message + assistant placeholder DIRECTLY into the
-    // chat-store, synchronously. The snapshot-on-switch path reads
-    // `messages` from the React closure, which lags behind setMessages
-    // by one render cycle (React batches). If the user clicks "+ new
-    // chat" in that gap, the snapshot writes stale messages (without
-    // the freshly-sent user message) to the store. Then the router
-    // takes over for the now-backgrounded session and only knows about
-    // assistant deltas — the user comes back and sees the assistant
-    // reply with no preceding user message. By writing both messages
-    // here, the store is at least as fresh as the panel and survives
-    // any closure staleness.
-    const sidNow = piSessionIdRef.current;
-    if (sidNow) {
-      const storeState = useChatStore.getState();
-      if (!storeState.sessions[sidNow]) {
-        storeState.actions.upsert({
-          id: sidNow,
-          title: "untitled",
-          preview: "",
-          status: "streaming",
-          messageCount: 0,
-          createdAt: Date.now(),
-          updatedAt: Date.now(),
-          pinned: false,
-          unread: false,
-        });
-      }
-      storeState.actions.appendMessage(sidNow, newUserMessage as any);
-      const currentTitle = useChatStore.getState().sessions[sidNow]?.title;
-      if (displayLabel && isPlaceholderConversationTitle(currentTitle)) {
-        storeState.actions.patch(sidNow, { title: displayLabel });
-      }
-      storeState.actions.appendMessage(sidNow, {
-        id: assistantMessageId,
-        role: "assistant",
-        content: "Processing...",
-        timestamp: Date.now(),
-        model: activePreset?.model,
-        provider: activePreset?.provider,
-      } as any);
-      storeState.actions.setStreaming(sidNow, {
-        streamingMessageId: assistantMessageId,
-        streamingText: "",
-        contentBlocks: [],
-        isStreaming: true,
-        isLoading: true,
-      });
-      // Bump the sort key — sending a message is the one user action
-      // that should pull a chat to the top of the sidebar. Pi-driven
-      // activity (text_delta, agent_end) does NOT bump this; the
-      // sidebar order is otherwise stable.
-      storeState.actions.patch(sidNow, { lastUserMessageAt: Date.now() });
-      // Clear the draft flag so the chat appears in the sidebar immediately.
-      // Without this, navigating away before the assistant responds leaves
-      // the session hidden (draft:true) even though there's a user message.
-      storeState.actions.patch(sidNow, { draft: false });
-    }
-
-    posthog.capture("chat_message_sent", {
-      provider: activePreset?.provider,
-      model: activePreset?.model,
-      has_images: outgoingImages.length > 0 || !!prefillFrameId,
-      has_context: !!prefillContext,
-      message_index: messages.filter((m) => m.role === "user").length,
-    });
-
-    // No timeout — Pi can run for minutes on long tasks (e.g. 30-day analysis
-    // with many tool calls). Process death is detected via pi_terminated event.
-    const timeoutId: ReturnType<typeof setTimeout> | null = null;
-
-    try {
-      // Collect images (pasted image + prefill frame)
-      // Pi expects ImageContent format: { type: "image", mimeType: "...", data: "..." }
-      const piImages: Array<{ type: string; mimeType: string; data: string }> = [];
-
-      if (prefillFrameId) {
-        try {
-          const response = await localFetch(`/frames/${prefillFrameId}`);
-          if (response.ok) {
-            const blob = await response.blob();
-            const arrayBuffer = await blob.arrayBuffer();
-            const base64 = btoa(
-              new Uint8Array(arrayBuffer).reduce((data, byte) => data + String.fromCharCode(byte), '')
-            );
-            const mimeType = blob.type || 'image/png';
-            piImages.push({
-              type: "image",
-              mimeType,
-              data: base64,
-            });
-          }
-        } catch (error) {
-          console.error("Failed to fetch frame image for Pi:", error);
-        }
-        setPrefillFrameId(null);
-      }
-
-      // Clear prefill context banner (was only cleared in non-Pi path)
-      if (prefillContext) {
-        // Prepend context to the user message so Pi sees it
-        const contextLabel = prefillSource === "timeline" ? "timeline selection" : "search";
-        userMessage = `[Context from ${contextLabel}: ${prefillContext}]\n\n${userMessage}`;
-        setPrefillContext(null);
-      }
-
-      for (const img of outgoingImages) {
-        const match = img.match(/^data:(image\/[^;]+);base64,(.+)$/);
-        if (match) {
-          piImages.push({
-            type: "image",
-            mimeType: match[1],
-            data: match[2],
-          });
-        }
-      }
-      if (shouldClearPastedImages) setPastedImages([]);
-
-      setMessages((prev) => [
-        ...prev,
-        { id: assistantMessageId, role: "assistant", content: "Processing...", timestamp: Date.now(), model: activePreset?.model, provider: activePreset?.provider },
-      ]);
-
-      // Always re-inject the recent conversation history into every prompt
-      // when the chat has prior turns (issue #3636).
-      //
-      // The previous contract gated injection on `piSessionSyncedRef.current`
-      // — a local boolean that tracked "we believe Pi has the conversation
-      // in its own in-memory session." The ref was reset on explicit Pi
-      // restarts (piStart paths), but Pi can also lose state silently —
-      // pi-agent runs context compaction by default (default settings:
-      // reserveTokens 16384, keepRecentTokens 20000), pi can crash and
-      // be auto-restarted before our termination handler observes the
-      // exit, and a queued / steer follow-up can race with a fresh
-      // sendPiMessage in ways the ref can't track. When the ref says
-      // "synced" but Pi has actually dropped everything, the next turn
-      // is sent as a bare user message — the model sees no prior context
-      // and answers as if the conversation just started. That's the
-      // user-visible symptom in issue #3636: "chat suddenly loses prior
-      // conversation context, but if I explicitly ask it to read the
-      // previous conversation, it can."
-      //
-      // The frontend's `messages` array is the durable source of truth
-      // (it's what gets persisted to disk on every save). Sending the
-      // last ~40 turns every time costs a small amount of tokens against
-      // the model's context window, but eliminates the entire class of
-      // "pi state silently diverged from messages" bugs. Pi appends the
-      // prompt verbatim to its own session; in the steady-state path the
-      // model sees a small amount of duplication between Pi's accumulated
-      // state and the injected block, which it handles fine. In the
-      // failure path (Pi just restarted, compacted, or never had this
-      // turn at all), the injected block IS the conversation and the
-      // model has what it needs.
-      //
-      // `piSessionSyncedRef` is kept around because other code paths
-      // (preset change, reauth, the conversation-load handler) still
-      // toggle it for diagnostics, but it no longer gates injection.
-      const promptMessage = withConversationHistory(userMessage, messages);
-      piSessionSyncedRef.current = true;
-
-      // E2E test hook — write to __e2ePiPromptCaptures when the recorder is installed
-      {
-        const g = window as any;
-        if (Array.isArray(g.__e2ePiPromptCaptures)) {
-          g.__e2ePiPromptCaptures.push({
-            sessionId: piSessionIdRef.current,
-            message: promptMessage,
-            at: Date.now(),
-          });
-        }
-      }
-
-      const providerPreflight = await preflightChatProvider(activePreset);
-      if (!providerPreflight.ok) {
-        piStreamingTextRef.current = "";
-        piMessageIdRef.current = null;
-        piContentBlocksRef.current = [];
-        setMessages((prev) =>
-          prev.map((m) => m.id === assistantMessageId
-            ? { ...m, content: providerPreflight.message, retryPrompt: userMessage }
-            : m)
-        );
-        if (sidNow) {
-          const storeState = useChatStore.getState();
-          storeState.actions.patchMessage(sidNow, assistantMessageId, (m: any) => ({
-            ...m,
-            content: providerPreflight.message,
-            retryPrompt: userMessage,
-          }));
-          storeState.actions.setStreaming(sidNow, {
-            streamingMessageId: null,
-            streamingText: "",
-            contentBlocks: [],
-            isLoading: false,
-            isStreaming: false,
-          });
-        }
-        forceQueueModeRef.current = false;
-        setIsLoading(false);
-        setIsStreaming(false);
-        return;
-      }
-
-      // Send prompt — abort/new_session now await completion, so no retry needed
-      let result = await commands.piPrompt(
-        piSessionIdRef.current,
-        promptMessage,
-        piImages.length > 0 ? piImages : null,
-        null,
-      );
-
-      // Race: user hit "+ NEW" before Pi finished registering the new session
-      // in the pool. Auto-spawn once and retry before surfacing the error.
-      if (result.status === "error" && result.error.includes("Pi not initialized")) {
-        console.log("[Pi] session not registered yet — auto-spawning and retrying");
-        try {
-          const home = await homeDir();
-          const dir = await join(home, ".screenpipe", "pi-chat");
-          const providerConfig = buildProviderConfig();
-          const startRes = await commands.piStart(
-            piSessionIdRef.current,
-            dir,
-            settings.user?.token ?? null,
-            providerConfig,
-          );
-          if (startRes.status === "ok" && startRes.data.running) {
-            setPiInfo(startRes.data);
-            piSessionSyncedRef.current = false;
-            if (providerConfig) {
-              setRunningConfigFromProviderConfig(providerConfig);
-            }
-            syncThinkingLevelAfterStart(piSessionIdRef.current);
-            result = await commands.piPrompt(
-              piSessionIdRef.current,
-              promptMessage,
-              piImages.length > 0 ? piImages : null,
-              null,
-            );
-          }
-        } catch (e) {
-          console.error("[Pi] auto-spawn retry failed", e);
-        }
-      }
-
-      if (result.status === "error") {
-        if (timeoutId) clearTimeout(timeoutId);
-        piMessageIdRef.current = null;
-        // Provide helpful error messages for common failures
-        const rawError = result.error;
-        let errorMsg: string;
-        let retryPrompt: string | undefined;
-        const providerError = buildProviderErrorMessage(rawError, activePreset);
-
-        if (rawError.includes("already processing")) {
-          errorMsg = "The AI was mid-response when your message arrived.";
-          retryPrompt = userMessage;
-        } else if (rawError.includes("Broken pipe") || rawError.includes("not running") || rawError.includes("has died") || rawError.includes("Pi not initialized")) {
-          const provider = activePreset?.provider;
-          errorMsg = provider === "native-ollama"
-            ? "Ollama isn't running. Start it with: `ollama serve`"
-            : "AI agent crashed — restarting automatically...";
-          retryPrompt = userMessage;
-        } else if (providerError) {
-          errorMsg = providerError;
-          retryPrompt = userMessage;
-        } else if (rawError.includes("not found")) {
-          errorMsg = `Model "${activePreset?.model}" not found. Check your AI preset in settings.`;
-        } else {
-          errorMsg = rawError;
-          retryPrompt = userMessage;
-        }
-        setMessages((prev) =>
-          prev.map((m) =>
-            m.id === assistantMessageId
-              ? { ...m, content: errorMsg, ...(retryPrompt ? { retryPrompt } : {}) }
-              : m
-          )
-        );
-        forceQueueModeRef.current = false;
-        setIsLoading(false);
-        setIsStreaming(false);
-      }
-    } catch (error) {
-      if (timeoutId) clearTimeout(timeoutId);
-      piMessageIdRef.current = null;
-      const rawError = error instanceof Error ? error.message : "Unknown error";
-      const providerError = buildProviderErrorMessage(rawError, activePreset);
-      setMessages((prev) =>
-        prev.map((m) =>
-          m.id === assistantMessageId
-            ? { ...m, content: providerError || `Error: ${rawError}` }
-            : m
-        )
-      );
-      forceQueueModeRef.current = false;
-      setIsLoading(false);
-      setIsStreaming(false);
-    }
-  }
-
-  async function executeSearchTool(args: Record<string, unknown>): Promise<string> {
-    const MAX_LIMIT = 10;
-
-    try {
-      const params = new URLSearchParams();
-      if (args.q) params.append("q", String(args.q));
-      if (args.content_type && args.content_type !== "all") {
-        params.append("content_type", String(args.content_type));
-      }
-
-      const requestedLimit = args.limit ? Number(args.limit) : 10;
-      const limit = Math.min(requestedLimit, MAX_LIMIT);
-      params.append("limit", String(limit));
-
-      if (args.start_time) params.append("start_time", String(args.start_time));
-      if (args.end_time) params.append("end_time", String(args.end_time));
-      if (args.app_name) params.append("app_name", String(args.app_name));
-      if (args.window_name) params.append("window_name", String(args.window_name));
-      if (args.speaker_name) params.append("speaker_name", String(args.speaker_name));
-
-      const controller = new AbortController();
-      const searchTimeoutId = setTimeout(() => controller.abort(), 120000);
-
-      const response = await localFetch(`/search?${params.toString()}`, {
-        signal: controller.signal,
-      });
-      clearTimeout(searchTimeoutId);
-
-      if (!response.ok) throw new Error(`Search failed: ${response.status}`);
-
-      const data = await response.json();
-
-      if (data.error) {
-        return `Search failed: ${data.error}`;
-      }
-
-      return formatSearchToolResults((data.data || []) as SearchToolResult[]);
-    } catch (error) {
-      return formatSearchToolError(error);
-    }
-  }
-
-  async function sendMessage(userMessage: string, displayLabel?: string, imageDataUrls?: string[]) {
-    if ((!canChat && !autoSendBypassRef.current) || (!activePreset && !autoSendBypassRef.current)) return;
-    const trimmed = userMessage.trim();
-    const outgoingImages = imageDataUrls ?? pastedImages;
-    const queuedDocs = attachedDocsRef.current;
-    if (!trimmed && outgoingImages.length === 0 && queuedDocs.length === 0) return;
-
-    // Fold any attached documents into the outgoing turn. The extracted
-    // text rides in `content` (what the model sees, kept for
-    // history/retries) while the bubble renders `displayContent` (the
-    // clean prompt) plus an attachment row above it (icon + name).
-    // The raw `<attached file: ...>` payload never reaches the renderer:
-    // when attachments are present the bubble's expand-chevron is
-    // suppressed (see ChatMessage / CollapsibleUserMessage).
-    let outgoingMessage = trimmed;
-    let outgoingDisplay = displayLabel;
-    const snapshotDocs = queuedDocs.length > 0 ? [...queuedDocs] : [];
-    if (queuedDocs.length > 0) {
-      const docText = docsToPromptText(queuedDocs);
-      outgoingMessage = [trimmed, docText].filter(Boolean).join("\n\n");
-      // Always set a clean displayContent when docs are attached.
-      // Without it, the bubble would render `outgoingMessage` directly
-      // — dumping the extracted PDF prose into the chat.
-      const cleanLabel = trimmed || `📎 ${queuedDocs.map((d) => d.name).join(", ")}`;
-      outgoingDisplay = displayLabel ?? cleanLabel;
-      pendingAttachmentsRef.current = queuedDocs.map((d) => ({
-        name: d.name,
-        ext: d.ext,
-        charCount: d.charCount,
-        truncated: d.truncated,
-      }));
-      setAttachedDocs([]);
-    }
-
-    // Restore the chips if the downstream send path threw. Mirrors the
-    // pastedImages restore-on-error contract in enqueuePiMessage/sendPiMessage:
-    // a failed dispatch must not silently swallow the user's attachments.
-    const restoreDocsOnError = (e: unknown) => {
-      if (snapshotDocs.length > 0) {
-        setAttachedDocs((prev) => prev.length === 0 ? snapshotDocs : prev);
-      }
-      throw e;
-    };
-
-    try {
-      const taskHint = snapshotDocs.length > 0 && trimmed.length > 0 && trimmed.length <= 2_000
-        ? trimmed
-        : undefined;
-      const largeContext = await externalizeLargeContextIfNeeded(
-        outgoingMessage,
-        piSessionIdRef.current,
-        taskHint,
-      );
-      if (largeContext) {
-        outgoingMessage = largeContext.prompt;
-        outgoingDisplay = outgoingDisplay ?? largeContext.displayLabel;
-        pendingAttachmentsRef.current = [
-          ...pendingAttachmentsRef.current,
-          largeContext.attachment,
-        ];
-        toast({
-          title: "large context saved as file",
-          description: "Pi will use local chunk files instead of sending the full text inline.",
-        });
-      }
-    } catch (e) {
-      toast({
-        title: "failed to save large context",
-        description: "The message was not sent because the pasted input is too large to send inline.",
-        variant: "destructive",
-      });
-      restoreDocsOnError(e);
-    }
-
-    // Guard the tiny gap between submit and React's loading state update.
-    // During this window, rapid Enter presses must queue (not start a second
-    // normal turn), otherwise user bubbles can drift.
-    if (forceQueueModeRef.current || sendDispatchInFlightRef.current || piMessageIdRef.current || isLoading || isStreaming) {
-      try {
-        return await enqueuePiMessage(outgoingMessage, outgoingDisplay, imageDataUrls);
-      } catch (e) {
-        restoreDocsOnError(e);
-      }
-    }
-
-    sendDispatchInFlightRef.current = true;
-    try {
-      // All providers route through Pi agent
-      return await sendPiMessage(outgoingMessage, outgoingDisplay, imageDataUrls);
-    } catch (e) {
-      restoreDocsOnError(e);
-    } finally {
-      sendDispatchInFlightRef.current = false;
-    }
-  }
-
-  const openConnectionSetup = useCallback((connectionId: string) => {
-    window.dispatchEvent(
-      new CustomEvent("open-settings", {
-        detail: {
-          section: "connections",
-          connectionId: connectionId === "connections" ? null : connectionId,
-        },
-      }),
-    );
-  }, []);
-
-  async function queueFollowUpMessage(userMessage: string, displayLabel?: string) {
-    if ((!canChat && !autoSendBypassRef.current) || (!activePreset && !autoSendBypassRef.current)) return;
-    return enqueuePiMessage(userMessage, displayLabel);
-  }
-
-  function setAssistantInterruptedState(activeAssistantId: string | null, interruptedBySteer: boolean) {
-    if (!activeAssistantId) return;
-    let changed = false;
-    let nextRows: Message[] | null = null;
-    setMessages((prev) => {
-      const next = prev.map((message) => {
-        if (
-          message.id !== activeAssistantId ||
-          message.role !== "assistant" ||
-          Boolean(message.interruptedBySteer) === interruptedBySteer
-        ) {
-          return message;
-        }
-        changed = true;
-        return { ...message, interruptedBySteer };
-      });
-      if (changed) nextRows = next;
-      return changed ? next : prev;
-    });
-    if (!changed || !nextRows) return;
-    void saveConversation(nextRows, {
-      refreshHistory: false,
-      syncActiveConversation: false,
-    });
-    const sidNow = piSessionIdRef.current;
-    if (sidNow) {
-      useChatStore.getState().actions.setMessages(sidNow, nextRows as any);
-    }
-  }
-
-  function markCurrentAssistantInterrupted() {
-    setAssistantInterruptedState(piMessageIdRef.current, true);
-  }
-
-  function clearCurrentAssistantInterrupted() {
-    setAssistantInterruptedState(piMessageIdRef.current, false);
-  }
-
-  function buildSteerPrompt(batch: PendingSteerBatchItem[]) {
-    const latest = batch[batch.length - 1];
-    if (!latest) return "";
-
-    const originalUserMessage = latest.originalUserMessage.trim();
-    const steerMessages = batch
-      .map((item, index) => `${index + 1}. ${item.content}`)
-      .join("\n");
-
-    return [
-      "The user sent steering messages while the previous assistant response was still running.",
-      "Treat them as live steering for that turn: they may refine the original request, replace it, or redirect to a new request.",
-      "Infer the user's intent from the original request and the steering messages. If a steering message is a complete request, answer that request directly.",
-      "Apply steering messages in order. If they conflict, the final steering message has highest priority.",
-      "Do not explain the steering mechanism unless the user asks about it.",
-      "",
-      "Original user request:",
-      originalUserMessage || "(unknown previous request)",
-      "",
-      "Steering messages:",
-      steerMessages,
-      "",
-      "Final steering message:",
-      latest.content,
-      "",
-      "Now answer according to the final steered intent.",
-    ].join("\n");
-  }
-
-  function clearPendingSteerTransportState(sessionId = piSessionIdRef.current) {
-    pendingNextPiUserIntentRef.current = null;
-    pendingNextPiUserDisplayRef.current = null;
-    optimisticSteerRef.current = null;
-    if (sessionId) {
-      pendingSteerBatchRef.current = pendingSteerBatchRef.current.filter((item) => item.sessionId !== sessionId);
-      turnIntentLedgerRef.current = turnIntentLedgerRef.current.filter((record) =>
-        record.sessionId !== sessionId ||
-        record.kind !== "steer" ||
-        Boolean(record.consumedAssistantId)
-      );
-    }
-  }
-
-  /**
-   * Extracts the pending steer batch for the given session, clears it from
-   * the ref, computes derived fields (prompt, preview, images), and sets
-   * the intent / display / optimistic refs needed by the message_start
-   * handler to recognise the steer echo. Returns null when there is no
-   * pending batch for the session.
-   */
-  function prepareSteerBatch(sessionId: string) {
-    const batch = pendingSteerBatchRef.current.filter(
-      (item) => item.sessionId === sessionId,
-    );
-    if (batch.length === 0) return null;
-    pendingSteerBatchRef.current = pendingSteerBatchRef.current.filter(
-      (item) => item.sessionId !== sessionId,
-    );
-
-    const latest = batch[batch.length - 1];
-    const prompt = buildSteerPrompt(batch);
-    const preview = queuedPreviewForText(latest.content);
-    const combinedImages = imageDataUrlsToPiImages(
-      batch.flatMap((item) => item.images),
-    );
-
-    // Remove earlier batch items' turn intents (only latest survives).
-    batch.slice(0, -1).forEach((item) => removeTurnIntent(item.turnIntentId));
-
-    // Set intent refs so the message_start handler recognises the
-    // steer echo and creates the assistant placeholder.
-    pendingNextPiUserIntentRef.current = "steer";
-    pendingNextPiUserDisplayRef.current = {
-      preview,
-      images: [...latest.images],
-      ...(latest.attachments?.length
-        ? { attachments: [...latest.attachments] }
-        : {}),
-      ...(latest.displayContent
-        ? { displayContent: latest.displayContent }
-        : {}),
-      optimisticUserId: latest.optimisticUserId,
-      turnIntentId: latest.turnIntentId,
-    };
-    optimisticSteerRef.current = {
-      id: latest.optimisticUserId,
-      content: prompt,
-      turnIntentId: latest.turnIntentId,
-    };
-    registerTurnIntent({
-      id: latest.turnIntentId,
-      sessionId,
-      kind: "steer",
-      content: prompt,
-      preview,
-      displayedUserId: latest.optimisticUserId,
-      createdAt: latest.createdAt,
-    });
-
-    return { batch, latest, prompt, preview, combinedImages };
-  }
-
-  async function flushPendingSteerBatch() {
-    const sessionId = piSessionIdRef.current;
-    if (!sessionId || pendingSteerFlushInFlightRef.current) return;
-
-    const prepared = prepareSteerBatch(sessionId);
-    if (!prepared) return;
-    pendingSteerFlushInFlightRef.current = true;
-
-    const { batch, latest, prompt, preview, combinedImages } = prepared;
-    const interruptedAssistantId = batch.find((item) => item.interruptedAssistantId)?.interruptedAssistantId ?? null;
-    const hasActiveAssistant = Boolean(piMessageIdRef.current);
-
-    const labelMarkers: Message[] = batch.slice(0, -1).map((item, index) => ({
-      id: `${item.turnIntentId}-label`,
-      role: "assistant",
-      content: "",
-      intent: "steer",
-      turnIntentId: item.turnIntentId,
-      timestamp: Date.now() + index,
-      model: activePreset?.model,
-      provider: activePreset?.provider,
-    }));
-    const labelMarkerIds = new Set(labelMarkers.map((marker) => marker.id));
-
-    let nextRowsAfterLabels: Message[] | null = null;
-    if (labelMarkers.length > 0) {
-      setMessages((prev) => {
-        const existingIds = new Set(prev.map((message) => message.id));
-        const markersToAppend = labelMarkers.filter((marker) => !existingIds.has(marker.id));
-        if (markersToAppend.length === 0) return prev;
-        const next = [...prev, ...markersToAppend];
-        nextRowsAfterLabels = next;
-        return next;
-      });
-      if (nextRowsAfterLabels) {
-        void saveConversation(nextRowsAfterLabels, {
-          refreshHistory: false,
-          syncActiveConversation: false,
-        });
-        useChatStore.getState().actions.setMessages(sessionId, nextRowsAfterLabels as any);
-      }
-    }
-
-    let precreatedSteerAssistantId: string | null = null;
-    if (hasActiveAssistant) {
-      const steerAssistantId = `${latest.turnIntentId}-assistant`;
-      precreatedSteerAssistantId = steerAssistantId;
-      const steerAssistantPlaceholder: Message = {
-        id: steerAssistantId,
-        role: "assistant",
-        content: "Processing...",
-        intent: "steer",
-        turnIntentId: latest.turnIntentId,
-        steeredResponse: true,
-        timestamp: Date.now(),
-        model: activePreset?.model,
-        provider: activePreset?.provider,
-      };
-      let nextRowsAfterAssistant: Message[] | null = null;
-      setMessages((prev) => {
-        if (prev.some((message) => message.id === steerAssistantId)) return prev;
-        const steerUserIndex = prev.findIndex((message) => message.id === latest.optimisticUserId);
-        const insertIndex = steerUserIndex >= 0 ? steerUserIndex + 1 : prev.length;
-        const next = [
-          ...prev.slice(0, insertIndex),
-          steerAssistantPlaceholder,
-          ...prev.slice(insertIndex),
-        ];
-        nextRowsAfterAssistant = next;
-        return next;
-      });
-      if (nextRowsAfterAssistant) {
-        void saveConversation(nextRowsAfterAssistant, {
-          refreshHistory: false,
-          syncActiveConversation: false,
-        });
-        useChatStore.getState().actions.setMessages(sessionId, nextRowsAfterAssistant as any);
-      }
-      markTurnIntentConsumed(latest.turnIntentId, steerAssistantId);
-      piMessageIdRef.current = steerAssistantId;
-      piStreamingTextRef.current = "";
-      piContentBlocksRef.current = [];
-      useChatStore.getState().actions.setStreaming(sessionId, {
-        streamingMessageId: steerAssistantId,
-        streamingText: "",
-        contentBlocks: [],
-        isStreaming: true,
-        isLoading: true,
-      });
-    }
-
-    lastUserMessageRef.current = latest.content;
-    setIsLoading(true);
-    setIsStreaming(true);
-
-    try {
-      const result = hasActiveAssistant
-        ? await commands.piSteer(
-            sessionId,
-            prompt,
-            combinedImages.length > 0 ? combinedImages : null,
-          )
-        : await commands.piPrompt(
-            sessionId,
-            prompt,
-            combinedImages.length > 0 ? combinedImages : null,
-            preview,
-          );
-
-      if (result.status !== "ok") {
-        pendingNextPiUserIntentRef.current = null;
-        pendingNextPiUserDisplayRef.current = null;
-        optimisticSteerRef.current = null;
-        removeTurnIntent(latest.turnIntentId);
-        setAssistantInterruptedState(interruptedAssistantId, false);
-        if (labelMarkerIds.size > 0) {
-          setMessages((prev) => prev.filter((message) => !labelMarkerIds.has(message.id)));
-        }
-        if (precreatedSteerAssistantId) {
-          setMessages((prev) => prev.filter((message) => message.id !== precreatedSteerAssistantId));
-          piMessageIdRef.current = null;
-          piStreamingTextRef.current = "";
-          piContentBlocksRef.current = [];
-        }
-        pendingSteerBatchRef.current = [...batch, ...pendingSteerBatchRef.current];
-        setIsLoading(false);
-        setIsStreaming(false);
-        toast({ title: "failed to send steered message", description: result.error, variant: "destructive" });
-      }
-    } catch (e) {
-      pendingNextPiUserIntentRef.current = null;
-      pendingNextPiUserDisplayRef.current = null;
-      optimisticSteerRef.current = null;
-      removeTurnIntent(latest.turnIntentId);
-      setAssistantInterruptedState(interruptedAssistantId, false);
-      if (labelMarkerIds.size > 0) {
-        setMessages((prev) => prev.filter((message) => !labelMarkerIds.has(message.id)));
-      }
-      if (precreatedSteerAssistantId) {
-        setMessages((prev) => prev.filter((message) => message.id !== precreatedSteerAssistantId));
-        piMessageIdRef.current = null;
-        piStreamingTextRef.current = "";
-        piContentBlocksRef.current = [];
-      }
-      pendingSteerBatchRef.current = [...batch, ...pendingSteerBatchRef.current];
-      setIsLoading(false);
-      setIsStreaming(false);
-      const description = e instanceof Error ? e.message : String(e);
-      toast({ title: "failed to send steered message", description, variant: "destructive" });
-    } finally {
-      pendingSteerFlushInFlightRef.current = false;
-    }
-  }
-
-  async function steerMessage(userMessage: string, displayLabel?: string, imageDataUrls?: string[]) {
-    const hasImages = imageDataUrls ? imageDataUrls.length > 0 : pastedImages.length > 0;
-    const trimmed = userMessage.trim();
-    if (!trimmed && !hasImages) return;
-
-    const hadActiveReply = isLoading || isStreaming || !!piMessageIdRef.current;
-    if (!hadActiveReply || !piInfo?.running) {
-      return sendPiMessage(trimmed, displayLabel, imageDataUrls);
-    }
-
-    posthog.capture("chat_message_steered", {
-      provider: activePreset?.provider,
-      model: activePreset?.model,
-      had_active_reply: hadActiveReply,
-      from_queue: !!imageDataUrls,
-    });
-
-    const outgoingImages = imageDataUrls ?? pastedImages;
-    const shouldClearPastedImages = imageDataUrls == null && pastedImages.length > 0;
-    const fallbackOriginalUserMessage = lastUserMessageRef.current;
-
-    piRateLimitRetries.current = 0;
-    lastUserMessageRef.current = trimmed;
-    const turnIntentId = `steer-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`;
-    const steerAttachments = consumePendingAttachments();
-    const optimisticUser: Message = {
-      id: turnIntentId,
-      role: "user",
-      content: trimmed,
-      ...(displayLabel ? { displayContent: displayLabel } : {}),
-      ...(outgoingImages.length ? { images: [...outgoingImages] } : {}),
-      ...(steerAttachments ? { attachments: steerAttachments } : {}),
-      intent: "steer",
-      turnIntentId,
-      timestamp: Date.now(),
-    };
-    markCurrentAssistantInterrupted();
-    const activeAssistantId = piMessageIdRef.current;
-    let originalUserMessage = fallbackOriginalUserMessage;
-    let nextRowsAfterOptimisticAppend: Message[] | null = null;
-    setMessages((prev) => {
-      const activeAssistantIndex = activeAssistantId
-        ? prev.findIndex((message) => message.id === activeAssistantId)
-        : -1;
-      if (activeAssistantIndex >= 0) {
-        for (let i = activeAssistantIndex - 1; i >= 0; i -= 1) {
-          const candidate = prev[i];
-          if (candidate?.role === "user" && candidate.intent !== "steer") {
-            originalUserMessage = candidate.content;
-            break;
-          }
-        }
-      }
-      if (activeAssistantIndex < 0) {
-        const next = [...prev, optimisticUser];
-        nextRowsAfterOptimisticAppend = next;
-        return next;
-      }
-
-      const activeAssistant = prev[activeAssistantIndex];
-      const hasVisibleAssistantContent = Boolean(
-        activeAssistant?.content &&
-        activeAssistant.content !== "Processing..."
-      ) || Boolean(activeAssistant?.contentBlocks?.length);
-      let insertIndex = hasVisibleAssistantContent
-        ? activeAssistantIndex + 1
-        : activeAssistantIndex;
-      while (
-        insertIndex < prev.length &&
-        prev[insertIndex]?.role === "user" &&
-        prev[insertIndex]?.intent === "steer"
-      ) {
-        insertIndex += 1;
-      }
-      const next = [
-        ...prev.slice(0, insertIndex),
-        optimisticUser,
-        ...prev.slice(insertIndex),
-      ];
-      nextRowsAfterOptimisticAppend = next;
-      return next;
-    });
-    if (nextRowsAfterOptimisticAppend) {
-      void saveConversation(nextRowsAfterOptimisticAppend, {
-        refreshHistory: false,
-        syncActiveConversation: false,
-      });
-    }
-    const sidNow = piSessionIdRef.current;
-    if (sidNow && nextRowsAfterOptimisticAppend) {
-      useChatStore.getState().actions.setMessages(sidNow, nextRowsAfterOptimisticAppend as any);
-    }
-    setInput("");
-    if (inputRef.current) inputRef.current.style.height = "auto";
-
-    if (shouldClearPastedImages) setPastedImages([]);
-
-    pendingSteerBatchRef.current = [
-      ...pendingSteerBatchRef.current,
-      {
-        turnIntentId,
-        sessionId: piSessionIdRef.current,
-        content: trimmed,
-        originalUserMessage,
-        interruptedAssistantId: activeAssistantId ?? undefined,
-        images: [...outgoingImages],
-        ...(steerAttachments ? { attachments: [...steerAttachments] } : {}),
-        ...(displayLabel ? { displayContent: displayLabel } : {}),
-        optimisticUserId: optimisticUser.id,
-        createdAt: Date.now(),
-      },
-    ];
-    if (hadActiveReply) {
-      const sid = piSessionIdRef.current;
-      if (sid) {
-        const prepared = prepareSteerBatch(sid);
-        if (!prepared) return;
-        const { batch, latest, prompt, combinedImages } = prepared;
-
-        piActiveStopRequestedRef.current = true;
-        const interruptedAssistantId =
-          latest.interruptedAssistantId ?? null;
-
-        // Send steer directly — no abort needed.
-        // send_immediate sets steer_in_flight in Rust, holding the
-        // drain loop until the steer turn's agent_start fires.
-        // If piSteer fails at the IPC layer, Pi never received the
-        // steer — revert is clean. Mid-stream failures surface as
-        // agent_end / response events, not IPC errors.
-        void commands
-          .piSteer(
-            sid,
-            prompt,
-            combinedImages.length > 0 ? combinedImages : null,
-          )
-          .then((result) => {
-            if (result.status !== "ok") {
-              console.warn("[steer] piSteer returned non-ok:", result);
-              revertFailedComposerSteer(
-                batch,
-                latest,
-                interruptedAssistantId,
-                result.error ?? "steer command rejected",
-              );
-            }
-          })
-          .catch((err: unknown) => {
-            console.warn("[steer] piSteer failed, reverting", err);
-            revertFailedComposerSteer(
-              batch,
-              latest,
-              interruptedAssistantId,
-              err instanceof Error ? err.message : String(err),
-            );
-          });
-      }
-      return;
-    }
-    if (!piMessageIdRef.current) {
-      void flushPendingSteerBatch();
-    }
-  }
-
-  /** Undo all side-effects of a failed composer steer. */
-  function revertFailedComposerSteer(
-    batch: typeof pendingSteerBatchRef.current,
-    latest: (typeof pendingSteerBatchRef.current)[number],
-    interruptedAssistantId: string | null,
-    errorDescription: string,
-  ) {
-    // Clear intent refs so message_start handler ignores the steer.
-    pendingNextPiUserIntentRef.current = null;
-    pendingNextPiUserDisplayRef.current = null;
-    optimisticSteerRef.current = null;
-    piActiveStopRequestedRef.current = false;
-    removeTurnIntent(latest.turnIntentId);
-
-    // Un-mark the assistant that was marked interrupted.
-    setAssistantInterruptedState(interruptedAssistantId, false);
-
-    // Remove only the optimistic steer user bubble inserted by steerMessage.
-    const optimisticId = latest.optimisticUserId;
-    setMessages((prev) =>
-      prev.filter(
-        (m) =>
-          !(
-            m.id === optimisticId &&
-            m.role === "user" &&
-            m.intent === "steer"
-          ),
-      ),
-    );
-
-    // Put the batch back so a retry or future steer can use it.
-    pendingSteerBatchRef.current = [
-      ...batch,
-      ...pendingSteerBatchRef.current,
-    ];
-
-    toast({
-      title: "failed to send steered message",
-      description: errorDescription,
-      variant: "destructive",
-    });
-  }
-
-  async function steerQueuedPrompt(prompt: PiQueuedPrompt) {
-    beginQueuedAction(prompt.id);
-    const queuedDisplay = takeQueuedDisplayById(currentQueueSessionId, prompt.id);
-    const existingTurnIntent = queuedDisplay?.turnIntentId
-      ? turnIntentLedgerRef.current.find((record) => record.sessionId === currentQueueSessionId && record.id === queuedDisplay.turnIntentId)
-      : turnIntentLedgerRef.current.find((record) => record.sessionId === currentQueueSessionId && record.queueId === prompt.id);
-    const turnIntentId = existingTurnIntent?.id ?? `queued-steer-${prompt.id}`;
-    const optimisticQueuedContent = existingTurnIntent?.kind === "steer"
-      ? existingTurnIntent.preview
-      : existingTurnIntent?.content ?? queuedDisplay?.preview ?? prompt.preview;
-    const optimisticQueuedUser: Message = {
-      id: turnIntentId,
-      role: "user",
-      content: optimisticQueuedContent,
-      ...(queuedDisplay?.displayContent ? { displayContent: queuedDisplay.displayContent } : {}),
-      ...(queuedDisplay?.images.length ? { images: [...queuedDisplay.images] } : {}),
-      ...(queuedDisplay?.attachments?.length ? { attachments: [...queuedDisplay.attachments] } : {}),
-      intent: "steer",
-      turnIntentId,
-      timestamp: Date.now(),
-    };
-    const interruptedAssistantBeforeSteer = piMessageIdRef.current;
-    try {
-      pendingNextPiUserIntentRef.current = "steer";
-      pendingNextPiUserDisplayRef.current = {
-        preview: existingTurnIntent?.preview ?? queuedDisplay?.preview ?? prompt.preview,
-        images: queuedDisplay?.images ? [...queuedDisplay.images] : [],
-        ...(queuedDisplay?.displayContent ? { displayContent: queuedDisplay.displayContent } : {}),
-        optimisticUserId: optimisticQueuedUser.id,
-        turnIntentId,
-      };
-      registerTurnIntent({
-        id: turnIntentId,
-        sessionId: currentQueueSessionId,
-        kind: "steer",
-        content: existingTurnIntent?.content ?? queuedDisplay?.preview ?? prompt.preview,
-        preview: existingTurnIntent?.preview ?? queuedDisplay?.preview ?? prompt.preview,
-        displayedUserId: optimisticQueuedUser.id,
-        queueId: prompt.id,
-        createdAt: existingTurnIntent?.createdAt ?? Date.now(),
-      });
-      markCurrentAssistantInterrupted();
-      let nextRowsAfterQueuedSteer: Message[] | null = null;
-      setMessages((prev) => {
-        if (prev.some((message) => message.turnIntentId === turnIntentId || message.id === optimisticQueuedUser.id)) {
-          return prev;
-        }
-        const next = [...prev, optimisticQueuedUser];
-        nextRowsAfterQueuedSteer = next;
-        return next;
-      });
-      if (nextRowsAfterQueuedSteer) {
-        void saveConversation(nextRowsAfterQueuedSteer, {
-          refreshHistory: false,
-          syncActiveConversation: false,
-        });
-        const sidNow = piSessionIdRef.current;
-        if (sidNow) {
-          useChatStore.getState().actions.setMessages(sidNow, nextRowsAfterQueuedSteer as any);
-        }
-      }
-      const result = await commands.piSteerQueued(piSessionIdRef.current, prompt.id);
-      if (result.status !== "ok") {
-        pendingNextPiUserIntentRef.current = null;
-        pendingNextPiUserDisplayRef.current = null;
-        removeTurnIntent(turnIntentId);
-        setMessages((prev) =>
-          prev.filter(
-            (m) =>
-              !(
-                m.id === optimisticQueuedUser.id &&
-                m.role === "user" &&
-                m.intent === "steer"
-              ),
-          ),
-        );
-        restoreQueuedDisplay(currentQueueSessionId, prompt.id, queuedDisplay);
-        setAssistantInterruptedState(interruptedAssistantBeforeSteer, false);
-        toast({ title: "failed to steer queued message", description: result.error, variant: "destructive" });
-        return;
-      }
-      if (!result.data) {
-        // Benign race: the queued prompt already left the queue and will
-        // render via the normal message_start path. Only remove the
-        // steer-specific optimistic user bubble — do not remove or disturb
-        // any transcript state that the normal message_start path may need.
-        pendingNextPiUserIntentRef.current = null;
-        pendingNextPiUserDisplayRef.current = null;
-        removeTurnIntent(turnIntentId);
-        setMessages((prev) =>
-          prev.filter(
-            (m) =>
-              !(
-                m.id === optimisticQueuedUser.id &&
-                m.role === "user" &&
-                m.intent === "steer"
-              ),
-          ),
-        );
-        restoreQueuedDisplay(currentQueueSessionId, prompt.id, queuedDisplay);
-        setAssistantInterruptedState(interruptedAssistantBeforeSteer, false);
-        toast({
-          title: "message already started",
-          description: "That follow-up has moved out of the queue.",
-        });
-        return;
-      }
-      if (currentQueueSessionId) {
-        removeQueuedPrompt(currentQueueSessionId, prompt.id);
-      }
-    } catch (e) {
-      pendingNextPiUserIntentRef.current = null;
-      pendingNextPiUserDisplayRef.current = null;
-      removeTurnIntent(turnIntentId);
-      setMessages((prev) =>
-        prev.filter(
-          (m) =>
-            !(
-              m.id === optimisticQueuedUser.id &&
-              m.role === "user" &&
-              m.intent === "steer"
-            ),
-        ),
-      );
-      restoreQueuedDisplay(currentQueueSessionId, prompt.id, queuedDisplay);
-      setAssistantInterruptedState(interruptedAssistantBeforeSteer, false);
-      toast({
-        title: "failed to steer queued message",
-        description: e instanceof Error ? e.message : String(e),
-        variant: "destructive",
-      });
-    } finally {
-      finishQueuedAction(prompt.id);
-    }
-  }
-
-  // Keep ref in sync so useEffect callbacks can call sendMessage
-  sendMessageRef.current = sendMessage;
-
-  const handleStop = async () => {
-    piActiveStopRequestedRef.current = true;
-    try {
-      await commands.piAbortActive(piSessionIdRef.current);
-    } catch (e) {
-      console.warn("[Pi] Failed to abort:", e);
-    }
-    piStreamingTextRef.current = "";
-    piMessageIdRef.current = null;
-    piContentBlocksRef.current = [];
-    if (abortControllerRef.current) {
-      abortControllerRef.current.abort();
-    }
-    setIsLoading(false);
-    setIsStreaming(false);
-  };
+  useChatPanelEffects({
+    inputRef,
+    showMentionDropdown,
+    isLoading,
+    isStreaming,
+    piActiveStopRequestedRef,
+    piSessionIdRef,
+    setIsLoading,
+    setIsStreaming,
+    appFilterOpen,
+    appItemsLength: appItems.length,
+    appsLoading,
+    tagsLoading,
+    refreshAppItems,
+    refreshTagItems,
+  });
+
+  usePipeWatchSession({
+    aiPresets: settings.aiPresets,
+    setActivePreset,
+    startPipeExecution,
+    loadConversationRef,
+    setIsStreaming,
+    setIsLoading,
+    piSessionIdRef,
+  });
+
+  const {
+    flushPendingSteerBatch,
+    handleStop,
+    openConnectionSetup,
+    queueFollowUpMessage,
+    sendMessage,
+    steerMessage,
+    steerQueuedPrompt,
+  } = usePiSendTransport({
+    abortControllerRef,
+    activePreset,
+    attachedDocsRef,
+    autoSendBypassRef,
+    buildProviderConfig,
+    canChat,
+    cancelStreamingMessageRender,
+    consumePendingAttachments,
+    currentQueueSessionId,
+    beginQueuedAction,
+    finishQueuedAction,
+    forceQueueModeRef,
+    input,
+    inputRef,
+    isLoading,
+    isStreaming,
+    lastUserMessageRef,
+    messages,
+    optimisticSteerRef,
+    pastedImages,
+    pendingNextPiUserDisplayRef,
+    pendingNextPiUserIntentRef,
+    pendingSteerBatchRef,
+    pendingSteerFlushInFlightRef,
+    piActiveStopRequestedRef,
+    piContentBlocksRef,
+    piCrashCountRef,
+    piInfo,
+    piMessageIdRef,
+    piPresetSwitchPromiseRef,
+    piRateLimitRetries,
+    piSessionIdRef,
+    piSessionSyncedRef,
+    piStartInFlightRef,
+    piStreamingTextRef,
+    prefillContext,
+    prefillFrameId,
+    prefillSource,
+    queuedPrompts,
+    registerTurnIntent,
+    markTurnIntentConsumed,
+    removeQueuedPrompt,
+    removeTurnIntent,
+    restartCurrentPiSession,
+    restoreQueuedDisplay,
+    saveConversation,
+    sendDispatchInFlightRef,
+    sendMessageRef,
+    setAttachedDocs,
+    setInput,
+    setIsLoading,
+    setIsStreaming,
+    setMessages,
+    setPastedImages,
+    setPiInfo,
+    setPiStarting,
+    setPrefillContext,
+    setPrefillFrameId,
+    setRunningConfigFromProviderConfig,
+    settings,
+    stagePendingAttachments,
+    syncThinkingLevelAfterStart,
+    takeQueuedDisplayById,
+    turnIntentLedgerRef,
+  });
+
+  usePiForegroundEvents({
+    activePreset,
+    buildProviderConfig,
+    cancelStreamingMessageRender,
+    clearPipeExecution,
+    consumeQueuedDisplayForStartedMessage,
+    findTurnIntentForUserStart,
+    flushPendingSteerBatch,
+    flushStreamingMessageRender,
+    forceQueueModeRef,
+    handleAgentEventDataRef,
+    handleInvalidatedAuthToken,
+    lastUserMessageRef,
+    markTurnIntentConsumed,
+    messages,
+    mountedRef,
+    optimisticSteerRef,
+    pendingNextPiUserDisplayRef,
+    pendingNextPiUserIntentRef,
+    pendingSteerBatchRef,
+    piActiveStopRequestedRef,
+    piContentBlocksRef,
+    piCrashCountRef,
+    piFirstCallRetried,
+    piIntentionallyStoppedPidsRef,
+    piStartInFlightRef,
+    piLastCrashRef,
+    piLastErrorRef,
+    piMessageIdRef,
+    piRateLimitRetries,
+    piRunningConfigRef,
+    piSessionIdRef,
+    piSessionSyncedRef,
+    piStoppedIntentionallyRef,
+    piStreamingTextRef,
+    piTerminationDedupRef,
+    piThinkingStartRef,
+    saveConversation,
+    scheduleStreamingMessageRender,
+    sessionActivityLastEmitAtRef,
+    sessionActivityLastSigRef,
+    setIsLoading,
+    setIsStreaming,
+    setMessages,
+    setPiInfo,
+    settings,
+    syncThinkingLevelAfterStart,
+    turnIntentTextValuesMatch,
+  });
+
+  const {
+    handleKeyDown,
+    handlePaste,
+    handleSubmit,
+  } = useChatComposerShellActions({
+    input,
+    setInput,
+    inputRef,
+    connectionChip,
+    setConnectionChip,
+    isMac,
+    isComposing,
+    mentions: {
+      isOpen: showMentionDropdown,
+      selectedIndex: selectedMentionIndex,
+      suggestions: filteredMentions,
+    },
+    mentionActions: {
+      close: () => setShowMentionDropdown(false),
+      selectNext: () => setSelectedMentionIndex((index) => Math.min(index + 1, filteredMentions.length - 1)),
+      selectPrevious: () => setSelectedMentionIndex((index) => Math.max(index - 1, 0)),
+      insert: insertMention,
+    },
+    pastedImages,
+    pendingDocsRef,
+    attachedDocsRef,
+    queuedPrompts,
+    steerShortcutInFlightRef,
+    isKnownConnectionId: (id) => INTEGRATION_ICON_KEYS.has(id),
+    handlePastedFiles,
+    attachPastedText,
+    sendMessage,
+    steerMessage,
+    steerQueuedPrompt,
+  });
 
   const activeSourceFooterMessageId =
     isLoading || isStreaming
@@ -3834,17 +989,6 @@ export function StandaloneChat({
         }}
       />
 
-      {/* Main content area with optional history sidebar — only used in
-          the floating overlay window. Home page hides this entirely
-          (`hideInlineHistory`) and the same list is rendered in the
-          main AppSidebar instead. */}
-
-      {/* Horizontal split: chat column on the left, BrowserSidebar on the
-          right. The browser panel is a sibling of the *whole* chat
-          column (messages + input), so when it opens it pushes both the
-          message scroller and the input bar — instead of the prior
-          structure where it sat next to messages only and the input bar
-          extended underneath it. */}
       <div className="flex-1 flex min-h-0" data-browser-panel-host>
       <div className="flex-1 flex flex-col min-w-0">
       <ChatMainPane
